### 2023-06-22 #5576 @kthyng
Hi @devarops and thanks for your submission! I am going to ping the editorial board now to see if they think it is in scope for JOSS, the main question being [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort), since this is close to the 1000 LOC cut off.
