### 2023-06-22 #5576 @devarops
Thank you, @kthyng. Are you counting only the 1166 lines of Python code?
