### 2023-07-19 #5576 @kthyng
Hi @devarops — a specific concern was brought up during scope review I'd like your commentary on:

> The algorithm implemented does not seem to be published anywhere, but is instead presented in the paper they submit to JOSS. In my view, this is a significant drawback, since in JOSS the software is reviewed, not the algorithm. I suggest they first submit their algorithm to a relevant journal and have it peer reviewed, and then submit the implementation to JOSS.

Does the algorithm described in your JOSS paper exist elsewhere, described in the literature, or is this a new algorithm?
