### 2023-07-19 #5576 @devarops
> Does the algorithm described in your JOSS paper exist elsewhere, described in the literature, or is this a new algorithm?

@kthyng:
The algorithm mentioned in our JOSS paper has already been introduced and discussed in the scientific literature. Specifically, it was originally presented in the following paper:

- Rojas-Mayoral, E., Méndez-Sánchez, F. A., Rojas-Mayoral, B., & Aguirre-Muñoz, A. (2019). Improving the efficiency of aerial rodent eradications by means of the numerical estimation of rodenticide density. SPREP LIBRARY, 47–50. https://library.sprep.org/sites/default/files/improving-efficiency-aerial-rodent-eradications-means-numerical-estimation-rodenticide-density.pdf

We have made the necessary updates to the manuscript to incorporate this information. Instead of providing a detailed algorithm description in our JOSS submission, we now refer the interested readers to the aforementioned reference for a comprehensive understanding of the algorithm.

Thank you for bringing this to our attention, and we apologize for any oversight in not explicitly referencing the prior publication.
