+++ CHANGELOG.md
@@ -1,0 +1,22 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+## [0.4.0] - 2023-11-14
+
+### Added
+
+### Fixed
+
+### Changed
+
+### Removed
+
+[unreleased]: https://github.com/IslasGECI/nerd/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/IslasGECI/nerd/compare/v0.3.0...v0.4.0
+[0.0.1]: https://github.com/IslasGECI/nerd/releases/tag/v0.0.1
