+++ .github/workflows/draft-pdf.yml
@@ -24,10 +24,3 @@ jobs:
         with:
           name: paper
           path: examples/paper.pdf
-      - name: Auto-commit paper.md
-        uses: stefanzweifel/git-auto-commit-action@master
-        with:
-          commit_message: 🤖 Apply automatic changes
-          file_pattern: 'examples/paper.md'
-          commit_user_name: Ciencia de Datos • GECI
-          commit_user_email: ciencia.datos@islas.org.mx
