### 2023-11-14 7ab7edc
🐋🪞 Mount examples in docker image
