### 2023-11-14 819bdb0
🐋🪞 Mount tests data in docker image
