+++ examples/calibration-demo.py
@@ -1,0 +1,235 @@
+#!/usr/bin/env python
+# coding: utf-8
+
+# In order to write equation \eqref{eq:densidadEsFlujoSobreProductoRapidezPorAncho} as a function of
+# the aperture diameter of the bait bucket, we express the mass flow rate of bait as a function of the
+# aperture diameter, $\dot{m}(d)$. To do this, the bait in the bucket was weighed and the time
+# required to empty the bucket was measured and repeated using several aperture diameters. Figure
+# \ref{fig:flujoDeApertura} shows the results from the calibration as well as the fitted model.
+
+# In[1]:
+
+
+import nerd
+import nerd.calibration
+import nerd.density_functions
+
+
+# In[2]:
+
+
+get_ipython().run_line_magic('matplotlib', 'inline')
+import matplotlib.pyplot as plt
+from matplotlib.patches import FancyArrowPatch
+import numpy as np
+import pandas as pd
+
+
+# ## Fit flow rate
+
+# In[3]:
+
+
+flow_data = pd.read_csv("/workdir/data/flow.csv")
+flow_data = flow_data[flow_data.bait_status == "new"][["aperture", "flow"]]
+
+
+# In[4]:
+
+
+aperture_diameters = flow_data.aperture.values
+flow_rates = flow_data.flow.values
+flow_rate_function = nerd.calibration.fit_flow_rate(aperture_diameters, flow_rates)
+
+
+# In[5]:
+
+
+x = np.linspace(min(aperture_diameters) - 10, max(aperture_diameters) + 10)
+y = flow_rate_function(x)
+fontsize = 15
+
+plt.plot(x, y)
+plt.plot(aperture_diameters, flow_rates, "o", markeredgecolor="k")
+plt.xlabel("Aperture diameter (mm)", size=fontsize)
+plt.ylabel("Mass flow rate (kg/s)", size=fontsize)
+plt.xlim(50, 100)
+plt.ylim(0, 3)
+plt.xticks(size=fontsize)
+plt.yticks(size=fontsize)
+plt.savefig("figures/calibration.png", dpi=300, transparent=True)
+
+
+# ![Flow rate $\dot{m}$ (kg/s) as a function of aperture diameter, $d$ (mm); each symbol represents a
+# calibration event and the black curve is the quadratic model fitted to the data.\label{fig:calibration}]("figures/calibration.png")
+
+# ## Swath width
+
+# In[6]:
+
+
+density_profile = pd.read_csv("/workdir/data/profile.csv")
+
+
+# In[7]:
+
+
+distance = density_profile.distance.values
+density_kg_per_ha = density_profile.density.values
+density = density_kg_per_ha / 1e4  # To convert densities to kg per square meter
+swath_width = nerd.calibration.get_swath_width(distance, density)
+
+
+# In[8]:
+
+
+size = 10
+fontsize = 15
+plt.plot(distance, density_kg_per_ha, "o")
+plt.xlabel("Position from flightpath (m)", size=fontsize)
+plt.ylabel("Bait Density (kg/ha)", size=fontsize)
+plt.xlim(-40, 40)
+plt.ylim(0, 6)
+plt.xticks(size=size)
+plt.yticks(size=size)
+plt.text(
+    -0.5,
+    3,
+    "Flightpath",
+    size=15,
+    color="k",
+    rotation=90,
+    bbox=dict(facecolor="w", edgecolor="none"),
+)
+plt.axvline(0, color="k")
+plt.savefig("figures/plots.png", dpi=300, transparent=True)
+
+
+# ![Aca va el pie de figura.\label{fig:plots}]("figures/plots.png")
+
+# ## Select best density function
+
+# In[10]:
+
+
[…]
