### 2023-11-21 a72cfca
📓👷 Stop generating paper from notebook
