+++ examples/paper.md
@@ -128,17 +128,15 @@ plt.xlim(50, 100)
 plt.ylim(0, 3)
 plt.xticks(size=fontsize)
 plt.yticks(size=fontsize)
-plt.savefig("figures/calibration.png", dpi=300, transparent=True)
+plt.savefig("examples/figures/calibration.png", dpi=300, transparent=True)
 ```
 
 
     
-![png](paper_files/paper_15_0.png)
-    
 
 
 ![Flow rate $\dot{m}$ (kg/s) as a function of aperture diameter, $d$ (mm); each symbol represents a
-calibration event and the black curve is the quadratic model fitted to the data.\label{fig:calibration}]("figures/calibration.png")
+calibration event and the black curve is the quadratic model fitted to the data.\label{fig:calibration}]("examples/figures/calibration.png")
 
 ## Swath width
 
@@ -176,16 +174,15 @@ plt.text(
     bbox=dict(facecolor="w", edgecolor="none"),
 )
 plt.axvline(0, color="k")
-plt.savefig("figures/plots.png", dpi=300, transparent=True)
+plt.savefig("examples/figures/plots.png", dpi=300, transparent=True)
 ```
 
 
     
-![png](paper_files/paper_20_0.png)
     
 
 
-![Aca va el pie de figura.\label{fig:plots}]("figures/plots.png")
+![Aca va el pie de figura.\label{fig:plots}]("examples/figures/plots.png")
 
 ## Select best density function
 
@@ -233,16 +230,15 @@ plt.plot(
 )
 plt.xlabel("Distance (m)", size=fontsize)
 plt.ylabel("Density (kg/m$^2$)", size=fontsize);
-plt.savefig("figures/density_profile.png")
+plt.savefig("examples/figures/density_profile.png")
 ```
 
 
     
-![png](paper_files/paper_25_0.png)
     
 
 
-![Aca va el pie de figura.\label{fig:density_profile}]("figures/density_profile.png")
+![Aca va el pie de figura.\label{fig:density_profile}]("examples/figures/density_profile.png")
 
 ## Calibration model
 
@@ -294,12 +290,11 @@ cbar.ax.set_ylabel("Density (kg/ha)", size=fontsize_labels)
 cbar.ax.tick_params(labelsize=fontsize_ticks)
 plt.axhline(18.0056, color="r", linewidth=2)
 plt.text(65, 18.6, "35 knot", size=fontsize_labels, color="k")
-plt.savefig("figures/contour_plot.png", dpi=300, transparent=True)
+plt.savefig("examples/figures/contour_plot.png", dpi=300, transparent=True)
 ```
 
 
     
-![png](paper_files/paper_29_0.png)
     
 
 
@@ -307,7 +302,7 @@ plt.savefig("figures/contour_plot.png", dpi=300, transparent=True)
 diameter $d$ (mm), and speed $s$ (km/hr). The horizontal axis shows the aperture diameter of the
 bait bucket and the vertical axis shows the helicopter's speed. The resulting bait density on the
 ground is shown in the second vertical color axis. $\sigma(d,s)= \frac{\dot{m}(d)}{s\cdot w}$.
-\label{fig:contour_plot}]("figures/contour_plot.png")
+\label{fig:contour_plot}]("examples/figures/contour_plot.png")
 
 The resulting three-dimensional model, $$\sigma(d,s)= \frac{\dot{m}(d)}{s\cdot w},$$ is shown in
 Figure \ref{fig:densidadDeAperturaYRapidez}. During the planning stage of an eradication campaign,
@@ -369,19 +364,14 @@ config_filepath = "/workdir/data/nerd_config.json"
 nerd_model = Nerd(config_filepath)
[…]
