### 2023-11-21 b5d155a
📈✍🏾 Add figure captions
