+++ examples/paper.md
@@ -135,8 +135,7 @@ plt.savefig("examples/figures/calibration.png", dpi=300, transparent=True)
     
 
 
-![Flow rate $\dot{m}$ (kg/s) as a function of aperture diameter, $d$ (mm); each symbol represents a
-calibration event and the black curve is the quadratic model fitted to the data.\label{fig:calibration}](figures/calibration.png)
+![Flow rate $\dot{m}$ (kg/s) as a function of the aperture diameter, $d$ (mm); each dot represents a calibration event and the blue line is the quadratic model fitted to the data.\label{fig:calibration}](figures/calibration.png)
 
 ## Swath width
 
@@ -182,7 +181,7 @@ plt.savefig("examples/figures/plots.png", dpi=300, transparent=True)
     
 
 
-![Bait density (kg/ha) as a function of the distance from the flight path (m). The green dots shows the measured density on the field during a calibration event. \label{fig:plots}](figures/plots.png)
+![Bait density $\sigma$ (kg/ha) as a function of the distance from the flight path (m). The green dots show the measured density on the ground after a calibration event. \label{fig:plots}](figures/plots.png)
 
 ## Select best density function
 
@@ -238,7 +237,7 @@ plt.savefig("examples/figures/density_profile.png")
     
 
 
-![Bait density (kg/ha) as a function of the distance from the flight path (m). The red line shows the fitted density model. \label{fig:density_profile}](figures/density_profile.png)
+![Bait density $\sigma$ (kg/ha) as a function of the distance from the flight path (m). The red line shows the fitted density model. \label{fig:density_profile}](figures/density_profile.png)
 
 ## Calibration model
 
@@ -298,8 +297,7 @@ plt.savefig("examples/figures/contour_plot.png", dpi=300, transparent=True)
     
 
 
-![Surface bait density $\sigma$ (kg/ha) as a function of aperturediameter $d$ (mm), and speed $s$ (km/hr). The horizontal axis shows the aperture diameter of thebait bucket and the vertical axis shows the helicopter's speed. The resulting bait density on the ground is shown in the second vertical color axis. $\sigma(d,s)= \frac{\dot{m}(d)}{s\cdot w}$.
-\label{fig:contour_plot}](figures/contour_plot.png)
+![Surface bait density $\sigma$ (kg/ha) on the color axis as a function of the aperture diameter $d$ (mm) of the bait bucket on the horizontal axis and the helicopter speed $s$ (km/hr) on the vertical axis. \label{fig:contour_plot}](figures/contour_plot.png)
 
 The resulting three-dimensional model, $$\sigma(d,s)= \frac{\dot{m}(d)}{s\cdot w},$$ is shown in Figure \ref{fig:contour_plot}. 
 
