### 2023-11-21 ba36f25
🔥📈 Remove redundant figure
