+++ examples/references.bib
@@ -115,7 +115,7 @@
   journal = {Biological Conservation}
 }
 
-@article{Mayoral-Rojas2019,
+@article{Rojas2019,
   ID = {42374},
   author = {Rojas-Mayoral, E. and Méndez-Sánchez, F.A. and Rojas-Mayoral, B. and Aguirre-Muñoz, A.},
   journal = {SPREP LIBRARY},
