### 2023-11-24 6fc4ea1
✍🏿 Update "Use case" section
