### 2023-11-24 06bfee8
✍🏿 Update "Demonstration using a configuration file" section
