+++ examples/references.bib
@@ -115,18 +115,6 @@
   journal = {Biological Conservation}
 }
 
-@article{Rojas2019,
-  ID = {42374},
-  author = {Rojas-Mayoral, E. and Méndez-Sánchez, F.A. and Rojas-Mayoral, B. and Aguirre-Muñoz, A.},
-  journal = {SPREP LIBRARY},
-  PEIN-ID = {403},
-  publisher = {International Union for Nature Conservation (IUCN)},
-  pages = {47--50},
-  title = {Improving the efficiency of aerial rodent eradications by means of the numerical estimation of rodenticide density},
-  url = {https://library.sprep.org/sites/default/files/improving-efficiency-aerial-rodent-eradications-means-numerical-estimation-rodenticide-density.pdf},
-  year  = {2019}
-}
-
 @article{Medina2011,
   doi = {10.1111/j.1365-2486.2011.02464.x},
   author = {Medina, F.M. and Bonnaud, E. and Vidal, E. and Tershy, B.R. and Zavaleta, E.S. and Donlan, C.J. and Keitt, B.S. and Le Corre, M. and Horwarth, S.V. and Nogales, M.},
@@ -159,6 +147,18 @@
   url = {http://www.jstor.org/stable/2463215},
   volume = {147},
   year = {1996}
+}
+
+@article{Rojas2019,
+  ID = {42374},
+  author = {Rojas-Mayoral, E. and Méndez-Sánchez, F.A. and Rojas-Mayoral, B. and Aguirre-Muñoz, A.},
+  journal = {SPREP LIBRARY},
+  PEIN-ID = {403},
+  publisher = {International Union for Nature Conservation (IUCN)},
+  pages = {47--50},
+  title = {Improving the efficiency of aerial rodent eradications by means of the numerical estimation of rodenticide density},
+  url = {https://web.archive.org/https://library.sprep.org/sites/default/files/improving-efficiency-aerial-rodent-eradications-means-numerical-estimation-rodenticide-density.pdf},
+  year  = {2019}
 }
 
 @article{Towns2006,
