+++ examples/references.bib
@@ -157,7 +157,7 @@
   publisher = {International Union for Nature Conservation (IUCN)},
   pages = {47--50},
   title = {Improving the efficiency of aerial rodent eradications by means of the numerical estimation of rodenticide density},
-  url = {https://web.archive.org/https://library.sprep.org/sites/default/files/improving-efficiency-aerial-rodent-eradications-means-numerical-estimation-rodenticide-density.pdf},
+  url = {https://is.gd/jeq2bx},
   year  = {2019}
 }
 
