+++ examples/paper.md
@@ -62,9 +62,9 @@ At its core, NERD is a probability density function that describes bait density 
 
 @Rojas2019 showed that the function $\sigma(x,y)$ to represent superficial bait density (kg/m$^2$) must comply with the following property:
 \begin{equation}
-\int_{-\frac{w}{2}}^{+\frac{w}{2}} \sigma(x)dx=\frac{\dot{m}}{s}
+\int_{-\frac{w}{2}}^{+\frac{w}{2}} \sigma(x)dx=\frac{\dot{m}}{s},
   \label{eq:integralDeDensidadEsflujoSobreRapidez}
-\end{equation},
+\end{equation}
 where $\dot{m}$ is the bait flow (kg/s), $s$ is the speed of the helicopter (m/s), and $w$ is the swath width (m).
 
 # Calibration
@@ -310,5 +310,5 @@ The NERD maps can identify bait gaps during eradication campaigns in the field, 
 NERD is an algorithm that describes bait density as a function of the aperture diameter and helicopter speed, based on past calibration experiments measuring rodenticide mass flow through a bait bucket.
 NERD can aid in planning aerial operations and during eradication campaigns by providing near real-time feedback and allowing for on-the-spot corrections. 
 The final product of NERD is a bait density map generated in a matter of seconds, allowing for the instant identification of bait gaps and ensuring efficient resource use.
-  
-References
+
+# References
