+++ examples/references.bib
@@ -152,12 +152,13 @@
 @article{Rojas2019,
   ID = {42374},
   author = {Rojas-Mayoral, E. and Méndez-Sánchez, F.A. and Rojas-Mayoral, B. and Aguirre-Muñoz, A.},
+  doi = {10.5281/zenodo.10214344},
   journal = {SPREP LIBRARY},
   PEIN-ID = {403},
   publisher = {International Union for Nature Conservation (IUCN)},
   pages = {47--50},
   title = {Improving the efficiency of aerial rodent eradications by means of the numerical estimation of rodenticide density},
-  url = {https://is.gd/jeq2bx},
+  url = {https://doi.org/10.5281/zenodo.10214344},
   year  = {2019}
 }
 
