### 2023-11-28 a56f2bc
🎨 Format citation
