+++ README.md
@@ -44,7 +44,7 @@ Lastly, explore the Jupyter notebooks at http://localhost:8080/
 
 ## References
 
-- [Improving the efficiency of aerial rodent eradications by means of the numerical estimation of rodenticide density](https://doi.org/10.5281/zenodo.10214344)
+- Rojas-Mayoral, E. (2019) «Improving the efficiency of aerial rodent eradications by means of the numerical estimation of rodenticide density». Island invasives: scaling up to meet the challenge, IUCN. [doi: 10.5281/zenodo.10214344.](https://doi.org/10.5281/zenodo.10214344)
 
 ---
 
