+++ examples/paper.md
@@ -30,7 +30,7 @@ Invasive rodents are present on approximately 90% of islands worldwide, seriousl
 Therefore, accurate bait density maps must be generated in real-time to maximize the efficiency of rodent eradication campaigns utilizing aerial dispersal methods. 
 Traditionally, conservationists rely on ground-level bait dispersion maps generated using Geographic Information Systems (GIS). 
 However, this approach is time-consuming and based on untested assumptions. 
-To improve the accuracy and efficiency of aerial operations, we developed NERD (Numerical Estimation of Rodenticide Density), an algorithm that performs highly precise calculations and provides mediate results. 
+To improve the accuracy and efficiency of aerial operations, we developed NERD (Numerical Estimation of Rodenticide Density), an algorithm that performs highly precise calculations and provides immediate results. 
 At its core, NERD  is a probability density function describing bait density on the ground as a function of the aperture diameter of the rodenticide bucket and helicopter speed. 
 We have confirmed the effectiveness of the model by successfully utilizing it in two island rodent eradication campaigns: mice eradication on San Benito Oeste (400 ha) in the Mexican Pacific and ship rat eradication on Cayo Centro (539 ha) of Banco Chinchorro in the Mexican Caribbean. 
 Notably, the Cayo Centro campaign is the largest rodent eradication ever conducted on a wet tropical island to date. We have proved the efficiency of NERD and its potential to reduce the overall cost of large-scale rodent eradication campaigns significantly.
@@ -38,7 +38,7 @@ Notably, the Cayo Centro campaign is the largest rodent eradication ever conduct
 # Introduction
 
 Invasive rodent species are incredibly deleterious to island ecosystems, especially those with levels of endemism or those without higher-order predators or predators occupying similar niches to the invasive rodents [@Meyers2000]. 
-Invasive rodent population dynamics are poorly understood on islands [Grant2015]. 
+Invasive rodent population dynamics are poorly understood on islands [@Harper2015]. 
 However, rodents on islands can cause native plant and animal species to decline rapidly and severely, even to extinction [@Medina2011; @Towns2006]. 
 The resultant losses are reflected in reduced biodiversity and, in many cases, invasive rodents becoming the dominant species. 
 In cases of severe rodent invasion, critical island ecosystem services are lost [@Towns2006]. 
@@ -69,7 +69,7 @@ where $\dot{m}$ is the bait flow (kg/s), $s$ is the speed of the helicopter (m/s
 
 # Calibration
 
-Assuming the density is independent of $x$ (i.e., $\sigma$ does not change along the swath width) and expressing the mass flow rate of the bait as a function of the aperture diameter), $\dot{m}(d)$, we obtain a two-parameter model:
+Assuming the density is independent of $x$ (i.e., $\sigma$ does not change along the swath width) and expressing the mass flow rate of the bait as a function of the aperture diameter, $\dot{m}(d)$, we obtain a two-parameter model:
 
 \begin{equation} \sigma(d,s)= \frac{\dot{m}(d)}{s\cdot w}. \end{equation}
 
@@ -77,15 +77,14 @@ We obtained the mass flow rate as a function of the aperture diameter of the bai
 We repeated this using several aperture diameters and a known initial mass.
 
 
-```python
+```python markdown-code-runner
 import nerd
 import nerd.calibration
[…]
