### 2023-11-28 59f3493
⚗️  Remove Markdown Code Runner to see if JOSS action can render the code
