### 2023-11-28 7ef8ff3
⚗️  Remove Markdown Code Runner to see if JOSS action can render the code
