### 2023-11-28 ade83df
👷 Sync paper with preprint if it runs
