### 2023-11-30 87f158f
🔨🐋  Install module in image
