### 2023-11-30 d5b4b2b
👷 Add build-image job as first job
