+++ .github/workflows/actions.yml
@@ -1,22 +1,67 @@
 name: Test and publish
 on: push
 jobs:
-  test-module:
-    name: Test Python module 🐍
+  build-image:
+    name: Build Docker image 🐋
     runs-on: ubuntu-latest
     steps:
     - name: Checkout repo
       uses: actions/checkout@main
-    - name: Build Docker image
-      run: docker build --tag islasgeci/nerd:latest --tag islasgeci/nerd:${GITHUB_SHA:0:4} .
+    - name: Set up Docker Buildx
+      uses: docker/setup-buildx-action@v3
+    - name: Build and export
+      uses: docker/build-push-action@v5
+      with:
+        context: .
+        tags: islasgeci/nerd:latest,islasgeci/nerd:${GITHUB_SHA:0:4}
+        outputs: type=docker,dest=/tmp/image.tar
+    - name: Upload artifact
+      uses: actions/upload-artifact@v3
+      with:
+        name: image
+        path: /tmp/image.tar
+
+  test-module:
+    name: Test Python module 🐍
+    needs: build-image
+    runs-on: ubuntu-latest
+    steps:
+    - name: Checkout repo
+      uses: actions/checkout@main
+    - name: Download artifact
+      uses: actions/download-artifact@v3
+      with:
+        name: image
+        path: /tmp
+    - name: Load image
+        run: |
+          docker load --input /tmp/image.tar
+          docker image ls -a          
     - name: Check format
       run: docker run islasgeci/nerd:${GITHUB_SHA:0:4} make check
     - name: Test
       run: docker run islasgeci/nerd:${GITHUB_SHA:0:4} make coverage
-    - name: Run preprint.md
-      run: docker run islasgeci/nerd:${GITHUB_SHA:0:4} markdown-code-runner --verbose examples/preprint.md
     - name: Run mutation testing
       run: docker run islasgeci/nerd:${GITHUB_SHA:0:4} make mutants
+
+  test-paper:
+    name: Test JOSS Paper
+    needs: build-image
+    runs-on: ubuntu-latest
+    steps:
+    - name: Checkout repo
+      uses: actions/checkout@main
+    - name: Download artifact
+      uses: actions/download-artifact@v3
+      with:
+        name: image
+        path: /tmp
+    - name: Load image
+        run: |
+          docker load --input /tmp/image.tar
+          docker image ls -a          
+    - name: Run preprint.md
+      run: docker run islasgeci/nerd:${GITHUB_SHA:0:4} markdown-code-runner --verbose examples/preprint.md
     - name: Remove markdown-code-runner
       run: |
         cp examples/preprint.md examples/paper.md
@@ -26,6 +71,13 @@ jobs:
       with:
         commit_message: 🤖 Sync paper with preprint
         file_pattern: 'examples/paper.md'
+        commit_user_name: Ciencia de Datos • GECI
+        commit_user_email: ciencia.datos@islas.org.mx
+    - name: Auto-commit figures
+      uses: stefanzweifel/git-auto-commit-action@master
+      with:
+        commit_message: 🤖 Autocommit figures
+        file_pattern: 'examples/figures/*.png'
         commit_user_name: Ciencia de Datos • GECI
         commit_user_email: ciencia.datos@islas.org.mx
 
@@ -41,23 +93,19 @@ jobs:
       run: docker run --volume ${PWD}/Makefile:/workdir/Makefile islasgeci/nerd_demo:${GITHUB_SHA:0:4} make check_notebook
     - name: Verify the notebooks run
[…]
