### 2023-11-30 f422de5
⚰️ Remove obsolete scripts
