+++ examples/proceedings.py
@@ -1,81 +1,0 @@
-#!/usr/bin/env python
-# coding: utf-8
-
-# ---
-# title: 'NERD: Numerical Estimation of Rodenticide Density'
-# tags:
-#   - aerial broadcast
-#   - bait bucket
-#   - Python
-#   - rodent eradication
-#   - rodenticide bait density
-# authors:
-#   - name: Evaristo Rojas-Mayoral
-#     orcid: 0000-0001-6812-9562
-#     corresponding: true
-#     affiliation: 1
-#   - name: Braulio Rojas-Mayoral
-#     orcid: 0000-0003-2358-2843
-#     affiliation: 1
-#   - name: Federico A. Méndez-Sánchez
-#     orcid: 0000-0002-3467-0008
-#     affiliation: 1
-# affiliations:
-#   - name: Grupo de Ecología y Conservación de Islas
-#     index: 1
-# bibliography: references.bib
-# 
-# ...
-
-# # Summary
-# 
-# Invasive rodents are present on approximately 90% of the world's islands and constitute one of the most serious threats to both endemic and native island species. The eradication of rodents is central to island conservation efforts and the aerial broadcast of rodenticide bait is the preferred dispersal method. To maximize the efficiency of rodent eradication campaigns utilizing aerial dispersal methods, the generation of accurate and real-time bait density maps are needed.
-# Traditionally, the creation of ground-level bait dispersion maps has relied on Geographic Information System (GIS), an approach that is time-consuming and based on untested assumptions. In order to improve accuracy and expedite the evaluation of aerial operations, we developed an algorithm called NERD: Numerical Estimation of Rodenticide Density, which performs calculations with high precision and provides immediate results. At its core, NERD is a probability density function describing the bait density on the ground as a function of the aperture diameter of the bait bucket and the helicopter speed. The effectiveness of the model was demonstrated through its successful utilization in two rodent eradication campaigns in Mexico: the mice eradication on San Benito Oeste Island (400 ha) in the Mexican Pacific, and the ship rat eradication on Cayo Centro Island (539 ha) from Banco Chinchorro, in the Mexican Caribbean. Notably, the latter campaign represents the largest rodent eradication on a wet tropical island to date. NERD's efficacy has been proven, and it has the potential to significantly reduce the overall cost of large-scale rodent eradication campaigns.
-
-# # Introduction
-# 
-# The effects of invasive rodent species on island ecosystems are incredibly deleterious, especially
-# on islands that present high levels of endemism and islands that have evolved in the absence of
-# predators occupying similar niches to the invasive rodent species or higher order predators
-# [@Meyers2000]. The population biology of invasive rodents on islands is still poorly understood [Grant2015], but the presence of rodents on islands can lead to the rapid decline, severe reduction and extinction of native plant and animal species [@Medina2011; @Towns2006]. The resultant losses are reflected in reduced biodiversity on the affected islands and in many cases, the emergence of the invasive rodent as the dominant species. In severe cases of rodent invasion, key island ecosystem services are lost [@Towns2006]. 
[…]
