### 2023-11-30 0e794ac
👷🐳 Mount example figures directory
