### 2023-11-30 bb56bc5
🐳👷 Copy files instead of mounting volumes
