### 2023-12-09 #5576 @ashleefv
I do see that one of my previous publications is listed among the relevant publications. However, I have no additional bandwidth for new commitments for the next 6 weeks. If it could wait until after Jan. 20 2024, I could do it. I need to say no if it should be handled before then.

Ashlee N. Ford Versypt, Ph.D. (she, her, hers)
Associate Professor, Department of Chemical and Biological Engineering
Affiliated Faculty, Department of Engineering Education, Department of Biomedical Engineering, and Institute for Artificial Intelligence and Data Science
University at Buffalo, The State University of New York
507 Furnas Hall, Buffalo, NY 14260
***@***.***
http://www.cbe.buffalo.edu/fordversypt
