### 2023-12-20 #5576 @ashleefv
ok. I will review in the last part of January.

Ashlee N. Ford Versypt, Ph.D. (she, her, hers)
Associate Professor, Department of Chemical and Biological Engineering
Affiliated Faculty, Department of Engineering Education, Department of Biomedical Engineering, and Institute for Artificial Intelligence and Data Science
University at Buffalo, The State University of New York
507 Furnas Hall, Buffalo, NY 14260
***@***.***
http://www.cbe.buffalo.edu/fordversypt
