### 2024-02-06 #5576 @graciellehigino
> ok. I will review in the last part of January. Ashlee N. Ford Versypt, Ph.D. (she, her, hers) Associate Professor, Department of Chemical and Biological Engineering Affiliated Faculty, Department of Engineering Education, Department of Biomedical Engineering, and Institute for Artificial Intelligence and Data Science University at Buffalo, The State University of New York 507 Furnas Hall, Buffalo, NY 14260 ***@***.*** http://www.cbe.buffalo.edu/fordversypt
> […](#)
> ________________________________ From: Gracielle Higino ***@***.***> Sent: Monday, December 18, 2023 8:43 PM To: openjournals/joss-reviews ***@***.***> Cc: Ashlee Ford Versypt ***@***.***>; Mention ***@***.***> Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: NERD: Numerical Estimation of Rodenticide Density (Issue #5576) Hi @ashleefv<https://github.com/ashleefv> thank you for your honest response! I believe we can wait until this date. Let us know then if you're still interested and we'll add you as a reviewer. — Reply to this email directly, view it on GitHub<[#5576 (comment)](https://github.com/openjournals/joss-reviews/issues/5576#issuecomment-1861966323)>, or unsubscribe<https://github.com/notifications/unsubscribe-auth/ADAP4NU4KWABTTIUTE4H4Z3YKDWNNAVCNFSM6AAAAAAZO275N6VHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMYTQNRRHE3DMMZSGM>. You are receiving this because you were mentioned.Message ID: ***@***.***>

Hi @ashleefv are you still available to review this? =D
