### 2024-05-03 #5576 @david-yannick
Hi Marcos,

I'd be more than happy to review

On Fri, May 3, 2024 at 2:46 PM Marcos Vital (LEQ-UFAL) <
***@***.***> wrote:

> Dear @david-yannick <https://github.com/david-yannick>, would you be
> willing to review this submission for JOSS?
>
> We carry out an open checklist-driven reviews here in GitHub issues, and
> follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
> This is still a pre-review issue, and if you accept we will have three
> reviewers and will move on to a new issue where the review will take place.
> If you are not able to review, please let us know if you have
> recommendations for other reviewers that could help with this submission.
>
