### 2024-05-09 #6736 @editorialbot
[REVIEW]: NERD: Numerical Estimation of Rodenticide Density
**Submitting author:** @devarops (<a href="http://orcid.org/0000-0001-6812-9562">Evaristo Rojas Mayoral</a>)
**Repository:** https://github.com/IslasGECI/nerd
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
