### 2024-05-15 #6736 @ashleefv
Regarding contribution and authorship: @devarops is the only of the three listed authors that appears to have contributed to the repository. There are 7 other repository contributors who are not listed as JOSS paper authors.
