### 2024-06-07 IslasGECI/nerd#53 @devarops
Review from `@ashleefv`
- [ ] Contribution and authorship: Has the submitting author (@devarops) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
We have added the [CONTRIBUITING.md](https://github.com/IslasGECI/nerd/blob/develop/CONTRIBUTING.md) file with this information.
- [ ] State of the field: Do the authors describe how this software compares to other commonly-used packages?

### `paper.md`

- [x] typo: Figure 2 caption density units should be kg/m^2 instead of kg/ha to match the y-axis units
- [x] Figure 3 caption also lists density as kg/ha. Is it more useful to stay in density kg/ha or kg/m^2 throughout the paper.md and the code? Consistency is desireable.
- [x] typo: "resorting" should be "restoring" in Introduction.
- [x] typo: Figure 1 caption "even" should be "event" and "line" should be "curve"
- [x] paper.md is inside the examples subfolder. It seems weird to include the code in the paper.md file.

> We have fixed the typos.
> We have removed the code from the `paper.md` file.

### `readme.md`

- [x] the link to where to explore the Jupyter notebooks does not work. The local host typically needs to be deployed by Jupyter on one's local machine.

### `examples/calibration-demo.ipynb`

- [x] but that disagrees with the calibration.demo.ipynb code and doesn't run on my computer either.
- [x] /workdir is not automatically set.
- [x] In [3]: I needed to change the file location to "../tests/data/flow.csv" for the csv file to be read.
- [x] In [6]: "../tests/data/profile.csv"
- [x] The text in the first section includes TeX commands for \eqref and Figure \ref cross-referencing that are not understood in the Jupyter notebook. They both render as ??? in the notebook.
- [x] plt.savefig("examples/figures/calibration.png", dpi=300, transparent=True)
- [x] The text after the figure saved as plots.png is in Spanish "Aca va el pie de figura" or "Here goes the figure caption" and does not render correctly in the .ipynb file.
- [x] The text after the figure saved as density_profile.png is in Spanish "Aca va el pie de figura" or "Here goes the figure caption" and does not render correctly in the .ipynb file.
- [x] The text after the figure saved as calibration.png is the caption for Figure 1 in the paper.md file. That Markdown text does not render correctly in the .ipynb file.
- [x] The text after the figure saved as contour_plot.png is the caption for Figure 3 in the paper.md file. That Markdown text does not render correctly in the .ipynb file.
- [x] The text in the last section includes TeX commands Figure \ref cross-referencing that is not understood in the Jupyter notebook and renders as ??? in the notebook. Recommended to move this text to the top of the notebook.
- [x] In [5]: "figures/calibration.png" works as the path. The paper.md file shows the following for this line:
- [x] In [8]: "figures/plots.png"
- [x] In [12]: "figures/density_profile.png"
- [x] In [15]: "figures/contour_plot.png"

### `examples/proceedings.ipynb`

- [x] This seems like a draft of paper.md and may not need to be in the repository anymore. I did not review this file or the preprint.md file as they are neither the paper.md or the main code or examples.

> We have removed `proceedings.ipynb` and `preprint.md`.

### `examples/tiling_demo.ipynb`

- [x] #56
- [x] The text before the Discussion section is in Spanish "Aca va el pie de figura" or "Here goes the figure caption" and does not render correctly in the .ipynb file.
- [x] The Discussion at the end of this notebook could be easily included as part of the introduction. Then Discussion and References sections would not be included. Currently References section is empty in this .ipynb file and is not needed.

### Throughout the nerd repository

- [x] I recommend replacing relative paths (dot notation) of imports

from . import
from .. import
```

with the specific absolute paths, e.g., in nerd\calibration\best_density_function.py

from nerd.calibration import get_rmse_from_function_array
from nerd import density_functions as df
```

PEP 8 standard explicitly recommends absolute imports: https://realpython.com/absolute-vs-relative-python-imports/
This recommendation applies to the following scripts:

- [x] nerd/init.py
- [x] nerd/mapping/init.py
- [x] nerd/io/init.py
- [x] nerd/density_functions/init.py
- [x] nerd/calibration/init.py
- [x] nerd/calibration/model.py
- [x] nerd/calibration/best_density_function.py
- [x] nerd/calibration/rmse.py

Your files nerd/io/import_data.py and nerd/mapping/tiling.py already use the recommended absolute paths for imports.

> We have replaced relative paths (dot notation) of imports with with the specific absolute paths.

---

This issue is part of a JOSS review:
- https://github.com/openjournals/joss-reviews/issues/6736#issuecomment-2113440341
