+++ examples/preprint.md
@@ -63,7 +63,7 @@ At its core, NERD is a probability density function that describes bait density 
 @Rojas2019 showed that the function $\sigma(x,y)$ to represent superficial bait density (kg/m$^2$) must comply with the following property:
 \begin{equation}
 \int_{-\frac{w}{2}}^{+\frac{w}{2}} \sigma(x)dx=\frac{\dot{m}}{s},
-  \label{eq:integralDeDensidadEsflujoSobreRapidez}
+  \label{eq:DensityIntegralIsFluxOverSpeed}
 \end{equation}
 where $\dot{m}$ is the bait flow (kg/s), $s$ is the speed of the helicopter (m/s), and $w$ is the swath width (m).
 
