+++ .github/config/.wordlist.txt
@@ -1,0 +1,24 @@
+Banco
+Braulio
+Cayo
+cdot
+Centro
+Chinchorro
+Conservación
+DensityIntegralIsFluxOverSpeed
+dx
+Ecología
+endemism
+eq
+Evaristo
+frac
+GIS
+Grupo
+Islas
+json
+Méndez
+Oeste
+orcid
+rodenticide
+situ
+Sánchez
