### 2024-06-13 6e00ba9
🔥 Remove preprint.md and code from paper.md
