+++ examples/paper.md
@@ -77,120 +77,16 @@ We obtained the mass flow rate as a function of the aperture diameter of the bai
 We repeated this using several aperture diameters and a known initial mass.
 
 
-```python
-import nerd
-import nerd.calibration
-import nerd.density_functions
-```
-
-
-```python
-import matplotlib.pyplot as plt
-from matplotlib.patches import FancyArrowPatch
-import numpy as np
-import pandas as pd
-```
-
 ## Fit flow rate
-
-
-```python
-flow_data = pd.read_csv("/workdir/tests/data/flow.csv")
-flow_data = flow_data[flow_data.bait_status == "new"][["aperture", "flow"]]
-```
-
-
-```python
-aperture_diameters = flow_data.aperture.values
-flow_rates = flow_data.flow.values
-flow_rate_function = nerd.calibration.fit_flow_rate(aperture_diameters, flow_rates)
-```
-
-
-```python
-x = np.linspace(min(aperture_diameters) - 10, max(aperture_diameters) + 10)
-y = flow_rate_function(x)
-fontsize = 15
-
-plt.plot(x, y)
-plt.plot(aperture_diameters, flow_rates, "o", markeredgecolor="k")
-plt.xlabel("Aperture diameter (mm)", size=fontsize)
-plt.ylabel("Mass flow rate (kg/s)", size=fontsize)
-plt.xlim(50, 100)
-plt.ylim(0, 3)
-plt.xticks(size=fontsize)
-plt.yticks(size=fontsize)
-plt.savefig("examples/figures/calibration.png", dpi=300, transparent=True)
-```
-
 
 
 ![Flow rate $\dot{m}$ (kg/s) as a function of the aperture diameter [$d$ (mm)]. Each dot represents a calibration event. The blue curve is the quadratic model fitted to the data.\label{fig:calibration}](figures/calibration.png)
 
 ## Swath width
 
-
-```python
-density_profile = pd.read_csv("/workdir/tests/data/profile.csv")
-```
-
-
-```python
-distance = density_profile.distance.values
-density_kg_per_ha = density_profile.density.values
-density = density_kg_per_ha / 1e4  # To convert densities to kg per square meter
-swath_width = nerd.calibration.get_swath_width(distance, density)
-```
-
-
 ## Select best density function
 
 
-```python
-aperture_diameter_data = 55  # milimetres
-helicopter_speed_data = 20.5778  # meters per second (40 knots)
-```
-
-
-```python
-density_function = nerd.calibration.get_best_density_function(
-    distance,
-    density,
-    aperture_diameter_data,
-    helicopter_speed_data,
-    swath_width,
-    flow_rate_function,
-)
-estimated_profile = nerd.solver(
-    aperture_diameter_data,
-    helicopter_speed_data,
-    swath_width,
-    density_function,
-    flow_rate_function,
-)
-```
-
-
-```python
-fontsize = 15
-x = np.linspace(min(distance), max(distance))
-y = estimated_profile(x)
-estimated_density = estimated_profile(distance)
-plt.figure(figsize=[11, 8])
-plt.plot(x, y, "r", label="estimated density")
-plt.vlines(distance, density, density + estimated_density - density, "k")
-plt.plot(
-    distance,
-    density,
-    "s",
-    color="orange",
-    markeredgecolor="black",
-    label="real density",
-)
-plt.xlabel("Distance (m)", size=fontsize)
-plt.ylabel("Density (kg/m$^2$)", size=fontsize);
-plt.savefig("examples/figures/density_profile.png")
-```
[…]
