+++ examples/preprint.md
@@ -1,307 +1,0 @@
----
-title: 'NERD: Numerical Estimation of Rodenticide Density'
-tags:
-  - aerial broadcast
-  - bait bucket
-  - Python
-  - rodent eradication
-  - rodenticide bait density
-authors:
-  - name: Evaristo Rojas-Mayoral
-    orcid: 0000-0001-6812-9562
-    corresponding: true
-    affiliation: 1
-  - name: Braulio Rojas-Mayoral
-    orcid: 0000-0003-2358-2843
-    affiliation: 1
-  - name: Federico A. Méndez-Sánchez
-    orcid: 0000-0002-3467-0008
-    affiliation: 1
-affiliations:
-  - name: Grupo de Ecología y Conservación de Islas
-    index: 1
-bibliography: references.bib
-
-...
-
-# Summary
-
-Invasive rodents are present on approximately 90% of islands worldwide, seriously threatening endemic and native island species and making rodent eradication central to island conservation. Aerial broadcast is the preferred dispersal method of rodenticide bait. 
-Therefore, accurate bait density maps must be generated in real-time to maximize the efficiency of rodent eradication campaigns utilizing aerial dispersal methods. 
-Traditionally, conservationists rely on ground-level bait dispersion maps generated using Geographic Information Systems (GIS). 
-However, this approach is time-consuming and based on untested assumptions. 
-To improve the accuracy and efficiency of aerial operations, we developed NERD (Numerical Estimation of Rodenticide Density), an algorithm that performs highly precise calculations and provides immediate results. 
-At its core, NERD  is a probability density function describing bait density on the ground as a function of the aperture diameter of the rodenticide bucket and helicopter speed. 
-We have confirmed the effectiveness of the model by successfully utilizing it in two island rodent eradication campaigns: mice eradication on San Benito Oeste (400 ha) in the Mexican Pacific and ship rat eradication on Cayo Centro (539 ha) of Banco Chinchorro in the Mexican Caribbean. 
-Notably, the Cayo Centro campaign is the largest rodent eradication ever conducted on a wet tropical island to date. We have proved the efficiency of NERD and its potential to reduce the overall cost of large-scale rodent eradication campaigns significantly.
-
-# Introduction
-
-Invasive rodent species are incredibly deleterious to island ecosystems, especially those with levels of endemism or those without higher-order predators or predators occupying similar niches to the invasive rodents [@Meyers2000]. 
-Invasive rodent population dynamics are poorly understood on islands [@Harper2015]. 
-However, rodents on islands can cause native plant and animal species to decline rapidly and severely, even to extinction [@Medina2011; @Towns2006]. 
-The resultant losses are reflected in reduced biodiversity and, in many cases, invasive rodents becoming the dominant species. 
-In cases of severe rodent invasion, critical island ecosystem services are lost [@Towns2006]. 
-The first step to restoring islands and recovering biodiversity is eradicating invasive rodents.
-Effective strategies to combat the detrimental effects of invasive rodent species on island ecosystems aim to minimize or eradicate rodent populations, facilitating the restoration of native species and crucial ecosystem processes. 
-The aerial broadcast of rodenticide is widely utilized and involves dispersing rodenticide bait pellets from helicopters over target areas. 
-Aerial broadcast effectively reduces rodent populations and has been successfully employed in numerous eradication campaigns [@Keitt2015].
-
-# Statement of need
-
[…]
