+++ .github/workflows/draft-pdf.yml
@@ -16,10 +16,6 @@ jobs:
         with:
           config_path: .github/config/.spellcheck.yml
           task_name: Markdown
-      - name: Remove markdown-code-runner
-        run: |
-          cp examples/preprint.md examples/paper.md
-          sed --in-place "s/python markdown-code-runner/python/g" examples/paper.md
       - name: Build proceedings PDF
         uses: openjournals/openjournals-draft-action@master
         with:

[…]
