+++ nerd/density_functions/__init__.py
@@ -1,1 +1,1 @@
-from .density_functions import uniform, triangular, normal  # noqa
+from nerd.density_functions.density_functions import uniform, triangular, normal  # noqa
