+++ nerd/io/__init__.py
@@ -1,3 +1,3 @@
-from .geo2utm import geo2utm  # noqa
-from .import_data import *  # noqa
-from .Nerd import *  # noqa
+from nerd.io.geo2utm import geo2utm  # noqa
+from nerd.io.import_data import *  # noqa
+from nerd.io.Nerd import *  # noqa
