+++ nerd/mapping/__init__.py
@@ -1,1 +1,1 @@
-from .tiling import *  # noqa
+from nerd.mapping.tiling import *  # noqa
