### 2024-06-18 28660ea
🚚📒 Rename examples folder and include Makefile and Dockerfile
