### 2024-06-18 1ab14dd
🚚🐋 Move Dockerfile.demo to example_notebooks
