+++ .github/workflows/draft-pdf.yml
@@ -20,9 +20,9 @@ jobs:
         uses: openjournals/openjournals-draft-action@master
         with:
           journal: joss
-          paper-path: examples/paper.md
+          paper-path: paper/paper.md
       - name: Upload
         uses: actions/upload-artifact@main
         with:
           name: paper
-          path: examples/paper.pdf
+          path: paper/paper.pdf
