### 2024-06-18 21f2209
🗣️ Add verbose option for pypi publish
