+++ .github/workflows/actions.yml
@@ -124,3 +124,4 @@ jobs:
       with:
         password: ${{ secrets.PYPI_TOKEN }}
         skip-existing: true
+        verbose: true
