### 2024-06-28 #6736 @jhabriel
I have finished my review of the NERD package. I also have some additional comments/suggestions, which are given below.

### Installation

* Installation through `pip` works fine. I also tested the docker version and it works as stated. However, I would recommend including the option of installing `NERD` from source (this will be interesting for someone who wants to work with the development version rather than the stable version). To allow for this, you will need to include a `requirements.txt` file with the package dependencies and the instructions on how to install from source.
* Please state in the `README` the Python versions that `NERD` supports.

### Citation

I would strongly recommend adding a `CITATION.cff` file. See https://citation-file-format.github.io/

### Community Guidelines

* I strongly recommend adding a `CODE_OF_CONDUCT.md` file. See https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-a-code-of-conduct-to-your-project
* I strongly recommend adding a `CONTRIBUTING.md` file, where there are clear guidelines on how one can contribute to the project. Specifically, are the formatting (flake8, isort, black), typing (mypy) , testing (pytest) requirements before opening a pull request.
* Please, add in the README.md how an user can establish contact in case of finding bugs / questions / feature requests etc.
* Please, give some hint about the purpose of the last link in the README.

### Examples

* I have encountered difficulty running the examples. There are a lot of errors and hyperlinks broken. I believe @ashleefv did a wonderful job at reviewing those, so I won’t go through those.

### Source code documentation and typing

* Several classes, functions and methods are lacking docstrings and type hints. They have to be fixed.

**Classes/Functions/Methods that are lacking docstrings**:

* nerd/calibration/best_density_function/get_density_function_array
* nerd/calibration/best_density_function/select_best_density_function_from_array
* nerd/calibration/rmse/get_rmse_from_function_array
* nerd/calibration/rmse/get_rmse
* nerd/io/Nerd
* nerd/io/Nerd.calculate_total_density
* nerd/io/geo2utm/geo2utm
* nerd/io/import_data/tracmap2csv
* nerd/io/import_data/import_tracmap
* nerd/io/import_data/import_calibration_data
* nerd/io/import_data/check_output_directory
* nerd/io/import_data/import_multifile_tracmap
* nerd/io/import_data/create_df_list
* nerd/io/import_data/select_parameters_by_index
* nerd/io/import_data/select_density_function
* nerd/mapping/tiling/slope_between_two_points
* nerd/mapping/tiling/orthogonal_slope
* nerd/mapping/tiling/safe_divition
* nerd/mapping/tiling/cell_edges_slopes
* nerd/mapping/tiling/calculate_cell_y_limits
* nerd/mapping/tiling/cell_y_coordinates
* nerd/mapping/tiling/calculate_cell_x_limits
* nerd/mapping/tiling/cell_x_coordinates
* nerd/mapping/tiling/generate_cell_from_coordinates
* nerd/mapping/tiling/calculate_cell_density_in_border
* nerd/mapping/tiling/density_in_tile
* nerd/mapping/tiling/is_inside_tile
* nerd/mapping/tiling/calculate_directions
* nerd/mapping/tiling/generate_tile_direction_arrays
* nerd/mapping/tiling/sign_of_direction
* nerd/mapping/tiling/reorder_end_tile
* nerd/mapping/tiling/check_directions
* nerd/mapping/tiling/generate_contours
* nerd/mapping/tiling/create_contour_polygon_list
* nerd/mapping/tiling/export_contour_list_as_shapefile
* nerd/mapping/tiling/generate_grid_density
* nerd/mapping/tiling/Tracks
* nerd/mapping/tiling/Tracks.y_coordinates
* nerd/mapping/tiling/Tracks.x_coordinates
* nerd/mapping/tiling/Tracks.bucket_logger
* nerd/mapping/tiling/Tracks.helicopter_speed
* nerd/mapping/tiling/Tracks.n_data
* nerd/mapping/tiling/calculate_total_density
* nerd/mapping/tiling/generate_uniform_density_array
* nerd/mapping/tiling/density_contours_intervals

**Classes/Functions/Methods that are lacking type hints**:

* nerd/calibration/best_density_function/get_density_function_array
nerd/calibration/best_density_function/select_best_density_function_from_array
* nerd/calibration/rmse/get_rmse_from_function_array
* nerd/calibration/rmse/get_rmse
* nerd/io/Nerd.calculate_total_density
* nerd/io/geo2utm/geo2utm
* nerd/io/import_data/import_tracmap
* nerd/io/import_data/import_calibration_data
* nerd/io/import_data/check_output_directory
* nerd/io/import_data/import_multifile_tracmap
* nerd/io/import_data/create_df_list
* nerd/io/import_data/select_parameters_by_index
* nerd/io/import_data/select_density_function
* nerd/mapping/tiling/slope_between_two_points
* nerd/mapping/tiling/orthogonal_slope
* nerd/mapping/tiling/safe_divition
* nerd/mapping/tiling/cell_edges_slopes
* nerd/mapping/tiling/calculate_cell_y_limits
* nerd/mapping/tiling/cell_y_coordinates
* nerd/mapping/tiling/calculate_cell_x_limits
* nerd/mapping/tiling/cell_x_coordinates
* nerd/mapping/tiling/generate_cell_from_coordinates
* nerd/mapping/tiling/calculate_cell_density_in_border
* nerd/mapping/tiling/density_in_tile
* nerd/mapping/tiling/is_inside_tile
* nerd/mapping/tiling/calculate_directions
* nerd/mapping/tiling/generate_tile_direction_arrays
* nerd/mapping/tiling/sign_of_direction
* nerd/mapping/tiling/reorder_end_tile
* nerd/mapping/tiling/check_directions
* nerd/mapping/tiling/generate_contours
* nerd/mapping/tiling/create_contour_polygon_list
* nerd/mapping/tiling/export_contour_list_as_shapefile
* nerd/mapping/tiling/generate_grid_density
* nerd/mapping/tiling/Tracks.y_coordinates
* nerd/mapping/tiling/Tracks.x_coordinates
* nerd/mapping/tiling/Tracks.bucket_logger
* nerd/mapping/tiling/Tracks.helicopter_speed
* nerd/mapping/tiling/Tracks.n_data
* nerd/mapping/tiling/calculate_total_density
* nerd/mapping/tiling/generate_uniform_density_array
* nerd/mapping/tiling/density_contours_intervals

### Testing
* The package is well tested. Good job.

### Paper
* The paper is lacking the State of Field. Basically, a paragraph where NERD is compared to alternative software (if any).
* I recommend avoiding the `plt.savefig` commands, and simply go for `plt.show()`.
* Could you add the city and country of *Grupo de Ecología y Conservación de Islas*?
* Figure 1: Possibly typo. Did you mean **event** instead of **even**?
* Figure 4: Please add label axes.
* Figure 4: Please add the colorbar with proper units.
* L81. I suggest removing *extremely*.
* L86. I suggest removing *incredibly*.
* L86. The word *notably* is used twice in the paragraph. Consider using a synonym.
