### 2024-07-05 #6736 @david-yannick
My review is complete, below are a few comments and questions:
-	Authors created NERD to address lack of means to produce bait density maps and identify gaps [48-49]. There is not a State of Field section, however, it mentioned that there is a lack in this area.
o	Are these exclusive to helicopter based releases and/or also be applicable to drones?
-	Perhaps included parameters to account for expected mean ground wind speed and direction? This would potentially impact how rodenticide is dispersed [68-69] ?
-	Add guidelines for 3rd party contribution
