### 2024-08-19 IslasGECI/nerd#54 @devarops
Review from `@jhabriel`
- [x] Installation through pip works fine. I also tested the docker version and it works as stated. However, I would recommend including the option of installing NERD from source (this will be interesting for someone who wants to work with the development version rather than the stable version). To allow for this, you will need to include a requirements.txt file with the package dependencies and the instructions on how to install from source.
- [x] Please state in the README the Python versions that NERD supports.

## Citation
- [x] I would strongly recommend adding a CITATION.cff file. See https://citation-file-format.github.io/

- [x] I strongly recommend adding a CODE_OF_CONDUCT.md file. See https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-a-code-of-conduct-to-your-project
- [x] I strongly recommend adding a CONTRIBUTING.md file, where there are clear guidelines on how one can contribute to the project. Specifically, are the formatting (flake8, isort, black), typing (mypy), testing (pytest) requirements before opening a pull request.
- [x] Please, add in the README.md how an user can establish contact in case of finding bugs / questions / feature requests etc.
- [x] Please, give some hint about the purpose of the last link in the README.

- [ ] I have encountered difficulty running the examples. There are a lot of errors and hyperlinks broken. I believe @ashleefv did a wonderful job at reviewing those, so I won’t go through those.

## Source code documentation and typing
- [x] Several classes, functions and methods are lacking docstrings and type hints. They have to be fixed.

### Classes/Functions/Methods that are lacking docstrings:
- [x] nerd/calibration/best_density_function/get_density_function_array
- [x] nerd/calibration/best_density_function/select_best_density_function_from_array
- [x] nerd/calibration/rmse/get_rmse_from_function_array
- [x] nerd/io/Nerd
- [x] nerd/io/Nerd.calculate_total_density
- [x] nerd/calibration/rmse/get_rmse
- [x] nerd/io/geo2utm/geo2utm
- [x] nerd/io/import_data/tracmap2csv
- [x] nerd/io/import_data/import_tracmap
- [x] nerd/io/import_data/import_calibration_data
- [x] nerd/io/import_data/check_output_directory
- [x] nerd/io/import_data/import_multifile_tracmap
- [x] nerd/io/import_data/create_df_list
- [x] nerd/io/import_data/select_parameters_by_index
- [x] nerd/io/import_data/select_density_function
- [x] nerd/mapping/tiling/slope_between_two_points
- [x] nerd/mapping/tiling/orthogonal_slope
- [x] nerd/mapping/tiling/safe_divition
- [x] nerd/mapping/tiling/cell_edges_slopes
- [x] nerd/mapping/tiling/calculate_cell_y_limits
- [x] nerd/mapping/tiling/cell_y_coordinates
- [x] nerd/mapping/tiling/calculate_cell_x_limits
- [x] nerd/mapping/tiling/cell_x_coordinates
- [x] nerd/mapping/tiling/generate_cell_from_coordinates
- [x] nerd/mapping/tiling/calculate_cell_density_in_border
- [x] nerd/mapping/tiling/density_in_tile
- [x] nerd/mapping/tiling/is_inside_tile
- [x] nerd/mapping/tiling/calculate_directions
- [x] nerd/mapping/tiling/generate_tile_direction_arrays
- [x] nerd/mapping/tiling/sign_of_direction
- [x] nerd/mapping/tiling/reorder_end_tile
- [x] nerd/mapping/tiling/check_directions
- [x] nerd/mapping/tiling/generate_contours
- [x] nerd/mapping/tiling/create_contour_polygon_list
- [x] nerd/mapping/tiling/export_contour_list_as_shapefile
- [x] nerd/mapping/tiling/generate_grid_density
- [x] nerd/mapping/tiling/Tracks
- [x] nerd/mapping/tiling/Tracks.y_coordinates
- [x] nerd/mapping/tiling/Tracks.x_coordinates
- [x] nerd/mapping/tiling/Tracks.bucket_logger
- [x] nerd/mapping/tiling/Tracks.helicopter_speed
- [x] nerd/mapping/tiling/Tracks.n_data
- [x] nerd/mapping/tiling/calculate_total_density
- [x] nerd/mapping/tiling/generate_uniform_density_array
- [x] nerd/mapping/tiling/density_contours_intervals

### Classes/Functions/Methods that are lacking type hints:
- [x] nerd/calibration/best_density_function/get_density_function_array
- [x] nerd/calibration/best_density_function/select_best_density_function_from_array
- [x] nerd/calibration/rmse/get_rmse_from_function_array
- [x] nerd/calibration/rmse/get_rmse
- [x] nerd/io/Nerd.calculate_total_density
- [x] nerd/io/geo2utm/geo2utm
- [x] nerd/io/import_data/import_tracmap
- [x] nerd/io/import_data/import_calibration_data
- [x] nerd/io/import_data/check_output_directory
- [x] nerd/io/import_data/import_multifile_tracmap
- [x] nerd/io/import_data/create_df_list
- [x] nerd/io/import_data/select_parameters_by_index
- [x] nerd/io/import_data/select_density_function
- [x] nerd/mapping/tiling/slope_between_two_points
- [x] nerd/mapping/tiling/orthogonal_slope
- [x] nerd/mapping/tiling/safe_divition
- [x] nerd/mapping/tiling/cell_edges_slopes
- [x] nerd/mapping/tiling/calculate_cell_y_limits
- [x] nerd/mapping/tiling/cell_y_coordinates
- [x] nerd/mapping/tiling/calculate_cell_x_limits
- [x] nerd/mapping/tiling/cell_x_coordinates
- [x] nerd/mapping/tiling/generate_cell_from_coordinates
- [x] nerd/mapping/tiling/calculate_cell_density_in_border
- [x] nerd/mapping/tiling/density_in_tile
- [x] nerd/mapping/tiling/is_inside_tile
- [x] nerd/mapping/tiling/calculate_directions
- [x] nerd/mapping/tiling/generate_tile_direction_arrays
- [x] nerd/mapping/tiling/sign_of_direction
- [x] nerd/mapping/tiling/reorder_end_tile
- [x] nerd/mapping/tiling/check_directions
- [x] nerd/mapping/tiling/generate_contours
- [x] nerd/mapping/tiling/create_contour_polygon_list
- [x] nerd/mapping/tiling/export_contour_list_as_shapefile
- [x] nerd/mapping/tiling/generate_grid_density
- [x] nerd/mapping/tiling/Tracks.y_coordinates
- [x] nerd/mapping/tiling/Tracks.x_coordinates
- [x] nerd/mapping/tiling/Tracks.bucket_logger
- [x] nerd/mapping/tiling/Tracks.helicopter_speed
- [x] nerd/mapping/tiling/Tracks.n_data
- [x] nerd/mapping/tiling/calculate_total_density
- [x] nerd/mapping/tiling/generate_uniform_density_array
- [x] nerd/mapping/tiling/density_contours_intervals

- [ ] The paper is lacking the State of Field. Basically, a paragraph where NERD is compared to alternative software (if any).
- [x] I recommend avoiding the `plt.savefig` commands, and simply go for `plt.show()`.
- [x] Figure 1: Possibly typo. Did you mean event instead of even?
- [x] Could you add the city and country of Grupo de Ecología y Conservación de Islas?
- [x] Figure 4: Please add label axes.
- [x] Figure 4: Please add the colorbar with proper units.
- [x] L81. I suggest removing extremely.
- [x] L86. I suggest removing incredibly.
- [x] L86. The word notably is used twice in the paragraph. Consider using a synonym.

---

This issue is part of a JOSS review:

- https://github.com/openjournals/joss-reviews/issues/6736#issuecomment-2197696888
