### 2024-08-19 IslasGECI/nerd#55 @devarops
Review from `@david-yannick`
- [x] Authors created NERD to address lack of means to produce bait density maps and identify gaps [48-49].
- [ ] Perhaps included parameters to account for expected mean ground wind speed and direction? This would potentially impact how rodenticide is dispersed [68-69] ?
- [x] There is not a State of Field section, however, it mentioned that there is a lack in this area. o Are these exclusive to helicopter based releases and/or also be applicable to drones?
- [x] Add guidelines for 3rd party contribution

---

This issue is part of a JOSS review:
- https://github.com/openjournals/joss-reviews/issues/6736#issuecomment-2211484760
