### 2024-08-22 365c905
📝 add install from source instructions
