### 2024-08-22 2f56a90
🚚✨ Change from setup to pyproject
