+++ nerd/__init__.py
@@ -1,1 +1,4 @@
-from nerd.solver import solver  # noqa
+"""The NERD algorithm performs calculations with increased accuracy, displaying results almost in real-time."""
+
+__version__ = "0.4.0"
+from .solver import solver  # noqa
