### 2024-08-22 e47ad1a
🐍 change python required version
