### 2024-08-22 72691fd
🔥 remove setup.py from format
