+++ Makefile
@@ -54,7 +54,6 @@ coverage: setup
 
 format:
 	black --line-length 100 ${module}
-	black --line-length 100 setup.py
 	black --line-length 100 tests
 
 init: setup tests
