### 2024-08-23 #6736 @devarops
@kthyng: Here's an update on our progress:

- We have incorporated 65% of the comments from `@ashleefv`.
- Regarding the comments from `@jhabriel`:
    - We added the citation file.
    - We specified the Python version in the README.
    - We included instructions for local installation from the source code.
    - Additionally, we incorporated two of the nine corrections suggested for the paper.

I'll provide a new update by the end of next week.
