+++ CONTRIBUTING.md
@@ -1,0 +1,47 @@
+# Contributing to [NERD](https://github.com/IslasGECI/nerd)
+We love your input! We want to make contributing to this project as easy and transparent as possible, whether it's:
+
+- Reporting a bug
+- Discussing the current state of the code
+- Submitting a fix
+- Proposing new features
+- Becoming a maintainer
+
+## We Develop with Github
+We use github to host code, to track issues and feature requests, as well as accept pull requests.
+
+## We Use [Github Flow](https://guides.github.com/introduction/flow/index.html), So All Code Changes Happen Through Pull Requests
+Pull requests are the best way to propose changes to the codebase (we use [Github Flow](https://guides.github.com/introduction/flow/index.html)). We actively welcome your pull requests:
+
+1. Fork the repo and create your branch from `develop`.
+2. If you've added code that should be tested, add tests.
+3. If you've changed APIs, update the documentation.
+4. Ensure the test suite passes.
+5. Make sure your code lints.
+6. Issue that pull request!
+
+## Any contributions you make will be under the AGPL-3.0 Software License
+In short, when you submit code changes, your submissions are understood to be under the same [AGPL-3.0 License](https://choosealicense.com/licenses/agpl-3.0/) that covers the project. Feel free to contact the maintainers if that's a concern.
+
+## Report bugs using Github's [issues](https://github.com/IslasGECI/nerd/issues)
+We use GitHub issues to track public bugs. Report a bug by [opening a new issue](https://github.com/IslasGECI/nerd/issues); it's that easy!
+
+## Write bug reports with detail, background, and sample code
+[This is an example from Craig Hockenberry](http://stackoverflow.com/q/12488905/180626) of a bug report.
+
+**Great Bug Reports** tend to have:
+
+- A quick summary and/or background
+- Steps to reproduce
+  - Be specific!
+  - Give sample code if you can. [Briandk stackoverflow question](http://stackoverflow.com/q/12488905/180626) includes sample code that *anyone* with a base R setup can run to reproduce what I was seeing
+- What you expected would happen
+- What actually happens
+- Notes (possibly including why you think this might be happening, or stuff you tried that didn't work)
+
+
+## [License](https://github.com/IslasGECI/nerd/blob/develop/LICENSE)
+By contributing, you agree that your contributions will be licensed under its AGPL-3.0 License.
+
+## References
+This document was adapted from [briandk's template](https://gist.github.com/briandk/3d2e8b3ec8daf5a27a62).
