### 2024-08-30 165dc16
💫 Add contributin guideline to readme 👯‍♂️👨‍❤️‍👨
