### 2024-08-30 d47edd7
💫 Add contributin guideline to readme 👯‍♂️👨‍❤️‍👨
