### 2024-08-30 #6736 @devarops
Hi @kthyng,

This week, we addressed all of `@jhabriel`'s observations regarding the community guidelines.

I'll keep you updated on our progress in addressing the rest of the reviewers' comments.
