### 2024-09-06 #6736 @devarops
Hi @kthyng,

This week, we addressed two of `@ashleefv`'s comments on `examples/tiling_demo.ipynb` and began working on some of `@jhabriel`'s comments regarding the missing docstrings.
