### 2024-09-09 dea5ae5
✅⌨️  Pass mypy
