+++ mypy.ini
@@ -1,0 +1,14 @@
+[mypy]
+
+[mypy-shapely.*]
+ignore_missing_imports = True
+[mypy-fiona.*]
+ignore_missing_imports = True
+[mypy-geojsoncontour.*]
+ignore_missing_imports = True
+[mypy-matplotlib.*]
+ignore_missing_imports = True
+[mypy-scipy.*]
+ignore_missing_imports = True
+[mypy-utm.*]
+ignore_missing_imports = True
