### 2024-09-09 eed167b
➕🐍 Install my[py]
