### 2024-09-09 0e21320
⌨️ ℹ️ ✅ pass mypy
