### 2024-09-09 7555136
➕📝⌨️ ℹ️  Add docstrings and tipying to Nerd class
