+++ nerd/io/geo2utm.py
@@ -2,7 +2,7 @@ import pandas as pd
 import utm
 
 
-def geo2utm(csv_filename):
+def _geo2utm(csv_filename):
     data = pd.read_csv(csv_filename)
     data["easting"], data["northing"], data["zone_number"], data["zone_letter"] = utm.from_latlon(
         data.Lat.values, data.Lon.values
