+++ nerd/calibration/__init__.py
@@ -1,5 +1,5 @@
 from nerd.calibration.fit_flow_rate import fit_flow_rate  # noqa
 from nerd.calibration.get_swath_width import get_swath_width  # noqa
 from nerd.calibration.model import model  # noqa
-from nerd.calibration.rmse import *  # noqa
+from nerd.calibration.rmse import _get_rmse  # noqa
 from nerd.calibration.best_density_function import *  # noqa
