+++ nerd/io/__init__.py
@@ -1,7 +1,7 @@
 from nerd.io.geo2utm import _geo2utm  # noqa
 from nerd.io.import_data import (  # noqa
     _tracmap2csv,
-    import_tracmap,
+    _import_tracmap,
     import_calibration_data,
     check_output_directory,
     import_multifile_tracmap,
