+++ tests/test_import_data.py
@@ -1,7 +1,7 @@
 import os
 import pandas as pd
 from nerd.density_functions.density_functions import normal
-from nerd.io import import_tracmap, _tracmap2csv, import_calibration_data, select_density_function
+from nerd.io import _import_tracmap, _tracmap2csv, import_calibration_data, select_density_function
 from pandas._testing import assert_frame_equal
 import numpy as np
 from numpy.testing import assert_array_almost_equal
@@ -27,7 +27,7 @@ def test_tracmap2csv():
 
 def test_import_tracmap_without_csvfilename():
     expected_utm_data = pd.read_csv("tests/data/expected_utm_data.csv")
-    obtained_utm_data = import_tracmap(tracmap_filename="tests/data/tracmap_sample_data.txt")
+    obtained_utm_data = _import_tracmap(tracmap_filename="tests/data/tracmap_sample_data.txt")
     assert_frame_equal(expected_utm_data, obtained_utm_data)
     os.remove("input_data.csv")
 
@@ -35,7 +35,7 @@ def test_import_tracmap_without_csvfilename():
 def test_import_tracmap_with_csvfilename():
     test_csv_filename = "test_csv_filename.csv"
     expected_utm_data = pd.read_csv("tests/data/expected_utm_data.csv")
-    obtained_utm_data = import_tracmap(
+    obtained_utm_data = _import_tracmap(
         tracmap_filename="tests/data/tracmap_sample_data.txt", csv_filename=test_csv_filename
     )
     assert os.path.isfile(test_csv_filename)
