+++ tests/test_import_data.py
@@ -1,7 +1,7 @@
 import os
 import pandas as pd
 from nerd.density_functions.density_functions import normal
-from nerd.io import _import_tracmap, _tracmap2csv, import_calibration_data, select_density_function
+from nerd.io import _import_tracmap, _tracmap2csv, _import_calibration_data, select_density_function
 from pandas._testing import assert_frame_equal
 import numpy as np
 from numpy.testing import assert_array_almost_equal
@@ -45,7 +45,7 @@ def test_import_tracmap_with_csvfilename():
 
 def test_import_calibration_data():
     expected_utm_data = "tests/data/expected_calibration_data.csv"
-    obtained_utm_data = import_calibration_data(expected_utm_data)
+    obtained_utm_data = _import_calibration_data(expected_utm_data)
     assert_array_almost_equal(
         obtained_utm_data, np.array([6.86905404e-04, -6.19897620e-02, 1.72530268e00])
     )
