+++ nerd/io/__init__.py
@@ -2,7 +2,7 @@ from nerd.io.geo2utm import _geo2utm  # noqa
 from nerd.io.import_data import (  # noqa
     _tracmap2csv,
     _import_tracmap,
-    import_calibration_data,
+    _import_calibration_data,
     check_output_directory,
     import_multifile_tracmap,
     create_df_list,
