+++ nerd/io/import_data.py
@@ -37,7 +37,7 @@ def _check_output_directory(output_path: str) -> None:
         os.mkdir(output_path)
 
 
-def import_multifile_tracmap(config_file, csv_filename):
+def _import_multifile_tracmap(config_file, csv_filename):
     df_list = create_df_list(config_file)
     df_concat = pd.concat(df_list)
     output_path = config_file.get("output_path")
