+++ nerd/io/__init__.py
@@ -5,7 +5,7 @@ from nerd.io.import_data import (  # noqa
     _import_calibration_data,
     _check_output_directory,
     _import_multifile_tracmap,
-    create_df_list,
+    _create_df_list,
     select_density_function,
     select_parameters_by_index,
 )
