+++ nerd/mapping/tiling.py
@@ -12,7 +12,7 @@ import numpy as np
 from typing import Tuple
 
 
-def slope_between_two_points(y2, y1, x2, x1):
+def _slope_between_two_points(y2, y1, x2, x1):
     return safe_divition(y2 - y1, x2 - x1)
 
 
@@ -28,12 +28,14 @@ def safe_divition(numerator, denominator):
 
 def cell_edges_slopes(x, y, node_index):
     start_slope = orthogonal_slope(
-        slope_between_two_points(
+        _slope_between_two_points(
             y[node_index + 1], y[node_index - 1], x[node_index + 1], x[node_index - 1]
         )
     )
     end_slope = orthogonal_slope(
-        slope_between_two_points(y[node_index + 2], y[node_index], x[node_index + 2], x[node_index])
+        _slope_between_two_points(
+            y[node_index + 2], y[node_index], x[node_index + 2], x[node_index]
+        )
     )
     return start_slope, end_slope
 
