+++ nerd/mapping/tiling.py
@@ -13,14 +13,14 @@ from typing import Tuple
 
 
 def _slope_between_two_points(y2: float, y1: float, x2: float, x1: float) -> float:
-    return safe_divition(y2 - y1, x2 - x1)
+    return _safe_divition(y2 - y1, x2 - x1)
 
 
 def _orthogonal_slope(slope: float) -> float:
-    return safe_divition(-1, slope)
+    return _safe_divition(-1, slope)
 
 
-def safe_divition(numerator, denominator):
+def _safe_divition(numerator, denominator):
     if denominator == 0:
         return np.inf
     return numerator / denominator
