+++ nerd/mapping/__init__.py
@@ -2,7 +2,7 @@ from nerd.mapping.tiling import (  # noqa
     _slope_between_two_points,
     _orthogonal_slope,
     _safe_divition,
-    cell_edges_slopes,
+    _cell_edges_slopes,
     calculate_cell_y_limits,
     cell_y_coordinates,
     calculate_cell_x_limits,
