+++ nerd/mapping/__init__.py
@@ -6,7 +6,7 @@ from nerd.mapping.tiling import (  # noqa
     _calculate_cell_y_limits,
     _cell_y_coordinates,
     _calculate_cell_x_limits,
-    cell_x_coordinates,
+    _cell_x_coordinates,
     generate_cell_from_coordinates,
     calculate_cell_density_in_border,
     density_in_tile,
