+++ nerd/mapping/__init__.py
@@ -10,7 +10,7 @@ from nerd.mapping.tiling import (  # noqa
     _generate_cell_from_coordinates,
     _calculate_cell_density_in_border,
     _density_in_tile,
-    is_inside_tile,
+    _is_inside_tile,
     calculate_directions,
     generate_tile_direction_arrays,
     sign_of_direction,
