+++ nerd/mapping/tiling.py
@@ -76,6 +76,10 @@ def _cell_x_coordinates(
 def _generate_cell_from_coordinates(
     x: list, y: list, node: int, stripe_width: float, spatial_resolution: int
 ) -> tuple:
+    return xx_generate_cell_from_coordinates(x, y, node, stripe_width)
+
+
+def xx_generate_cell_from_coordinates(x: list, y: list, node: int, stripe_width: float) -> tuple:
     r = stripe_width / 2
     start_orthogonal_slope, end_orthogonal_slope = _cell_edges_slopes(x, y, node)
     x_rect = _cell_x_coordinates(r, start_orthogonal_slope, end_orthogonal_slope, x, node)
