+++ nerd/mapping/__init__.py
@@ -8,6 +8,7 @@ from nerd.mapping.tiling import (  # noqa
     _calculate_cell_x_limits,
     _cell_x_coordinates,
     _generate_cell_from_coordinates,
+    xx_generate_cell_from_coordinates,
     _calculate_cell_density_in_border,
     _density_in_tile,
     _is_inside_tile,
