+++ nerd/mapping/__init__.py
@@ -11,7 +11,7 @@ from nerd.mapping.tiling import (  # noqa
     _calculate_cell_density_in_border,
     _density_in_tile,
     _is_inside_tile,
-    calculate_directions,
+    _calculate_directions,
     generate_tile_direction_arrays,
     sign_of_direction,
     reorder_end_tile,
