+++ nerd/mapping/tiling.py
@@ -110,12 +110,12 @@ def _is_inside_tile(x_rect: list, y_rect: list, points: np.ndarray) -> List[bool
 
 
 def _calculate_directions(x_rect: list, y_rect: list) -> float:
-    u, v = generate_tile_direction_arrays(x_rect, y_rect)
+    u, v = _generate_tile_direction_arrays(x_rect, y_rect)
     theta1 = sign_of_direction(u, v)
     return theta1
 
 
-def generate_tile_direction_arrays(x_rect, y_rect):
+def _generate_tile_direction_arrays(x_rect, y_rect):
     u = np.array([x_rect[1] - x_rect[0], y_rect[1] - y_rect[0]])
     v = np.array([x_rect[2] - x_rect[3], y_rect[2] - y_rect[3]])
     return u, v
