+++ nerd/mapping/tiling.py
@@ -111,7 +111,7 @@ def _is_inside_tile(x_rect: list, y_rect: list, points: np.ndarray) -> List[bool
 
 def _calculate_directions(x_rect: list, y_rect: list) -> float:
     u, v = _generate_tile_direction_arrays(x_rect, y_rect)
-    theta1 = sign_of_direction(u, v)
+    theta1 = _sign_of_direction(u, v)
     return theta1
 
 
@@ -121,7 +121,7 @@ def _generate_tile_direction_arrays(x_rect: list, y_rect: list) -> Tuple:
     return u, v
 
 
-def sign_of_direction(u, v):
+def _sign_of_direction(u, v):
     inner = np.inner(u, v)
     norms = np.linalg.norm(u) * np.linalg.norm(v)
     return np.arccos(np.clip(inner / norms, -1.0, 1.0))
