+++ nerd/mapping/tiling.py
@@ -151,7 +151,7 @@ def _generate_contours(
     return contour, dict(zip(contour.collections, contour.levels))
 
 
-def create_contour_polygon_list(contour, contour_dict):
+def _create_contour_polygon_list(contour, contour_dict):
     # Original code in https://github.com/chrishavlin/learning_shapefiles/blob/master/src/contourf_to_shp.py
     PolyList = []
     for col in contour.collections:
