+++ nerd/mapping/__init__.py
@@ -19,7 +19,7 @@ from nerd.mapping.tiling import (  # noqa
     _generate_contours,
     _create_contour_polygon_list,
     _export_contour_list_as_shapefile,
-    generate_grid_density,
+    _generate_grid_density,
     Tracks,
     calculate_total_density,
     generate_uniform_density_array,
