+++ nerd/mapping/__init__.py
@@ -20,7 +20,7 @@ from nerd.mapping.tiling import (  # noqa
     _create_contour_polygon_list,
     _export_contour_list_as_shapefile,
     _generate_grid_density,
-    Tracks,
+    _Tracks,
     calculate_total_density,
     generate_uniform_density_array,
     density_contours_intervals,
