+++ tests/test_mapping.py
@@ -52,7 +52,7 @@ class TestMapping(TestCase):
         self._x_coordinates = np.linspace(0, 10, 10)
         self._y_coordinates = np.linspace(0, 10, 10)
         self.helicopter_speed = [20, 21, 20, 18, 17, 15, 15, 15, 15, 15, 10, 5]
-        self.bucket_logger = np.array([1, 1, 1, 1, 0, 0, 0, 1, 1, 1])
+        self._bucket_logger = np.array([1, 1, 1, 1, 0, 0, 0, 1, 1, 1])
         self.total_density_reshaped = np.eye(10, 10)
         self.aperture_diameter = 90
         self.density_function = normal
@@ -93,7 +93,7 @@ class TestMapping(TestCase):
             {
                 "easting": self._x_coordinates,
                 "northing": self._y_coordinates,
-                "Logging_on": self.bucket_logger,
+                "Logging_on": self._bucket_logger,
                 "Speed": self.helicopter_speed[:10],
             }
         )
