+++ tests/test_mapping.py
@@ -330,7 +330,7 @@ class TestMapping(TestCase):
 
     def test_calculate_total_density_2(self):
         spatial_resolution = 2
-        x_grid_obtained, y_grid_obtained, total_density_grid_obtained = _calculate_total_density(
+        _, _, total_density_grid_obtained = _calculate_total_density(
             self.trackmap_data,
             self.config_json,
             spatial_resolution,
