+++ CHANGELOG.md
@@ -7,7 +7,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.4.0] - 2023-11-14
 
 ### Added
 
@@ -16,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 ### Removed
+
+## [0.4.1] - 2024-10-04
+
+## [0.4.0] - 2023-11-14
 
 [unreleased]: https://github.com/IslasGECI/nerd/compare/v0.4.0...HEAD
 [0.4.0]: https://github.com/IslasGECI/nerd/compare/v0.3.0...v0.4.0
