+++ CITATION.cff
@@ -11,4 +11,4 @@ authors:
     given-names: Federico A.
     orcid: https://orcid.org/0000-0002-3467-0008
 title: "NERD: Numerical Estimation of Rodenticide Density. Journal of Open Source Software"
-version: 0.4.0
+version: 0.4.1
