### 2024-10-16 0ce46b2
✅✍🏽 add Ensenada to spellcheck
