+++ .github/config/.wordlist.txt
@@ -9,6 +9,7 @@ DensityIntegralIsFluxOverSpeed
 dx
 Ecología
 endemism
+Ensenada
 eq
 Evaristo
 frac
@@ -20,5 +21,5 @@ Méndez
 Oeste
 orcid
 rodenticide
-situ
 Sánchez
+situ
