### 2024-10-16 bbfed33
🇲🇽 Add city and country in affiliation
