+++ CONTRIBUTING.md
@@ -14,11 +14,18 @@ We use github to host code, to track issues and feature requests, as well as acc
 Pull requests are the best way to propose changes to the codebase (we use [Github Flow](https://guides.github.com/introduction/flow/index.html)). We actively welcome your pull requests:
 
 1. Fork the repo and create your branch from `develop`.
-2. If you've added code that should be tested, add tests.
-3. If you've changed APIs, update the documentation.
-4. Ensure the test suite passes.
-5. Make sure your code lints.
-6. Issue that pull request!
+1. Make sure your changes comply with the following checklist:
+    - [ ] **Comments**: Are the comments clear and useful?
+    - [ ] **Complexity**: Could the code be made simpler? Would another developer be able to easily understand and use this code when they come across it in the future?
+    - [ ] **Design**: Is the code well-designed and appropriate for your system?
+    - [ ] **Documentation**: Did the developer also update relevant documentation?
+    - [ ] **Functionality**: Does the code behave as the author likely intended? Is the way the code behaves good for its users?
+    - [ ] **Interface**: Any interface changes are sensible and look good.
+    - [ ] **Naming**: Did the developer choose clear names for variables, classes, methods, etc.?
+    - [ ] **Style**: Does the code follow our style guides?
+    - [ ] **Tests**: Does the code have correct and well-designed automated tests?
+    - [ ] **YAGNI**: The developer is not implementing things they might need in the future but don’t know they need now.
+1. Create that pull request!
 
 ## Any contributions you make will be under the AGPL-3.0 Software License
 In short, when you submit code changes, your submissions are understood to be under the same [AGPL-3.0 License](https://choosealicense.com/licenses/agpl-3.0/) that covers the project. Feel free to contact the maintainers if that's a concern.
