### 2025-01-16 8eeaad1
✅🎨 pass mypy
