+++ nerd/mapping/tiling.py
@@ -9,7 +9,7 @@ from tqdm import tqdm
 import fiona
 import matplotlib.pyplot as plt
 import numpy as np
-from typing import Tuple, List, Dict
+from typing import Tuple, List, Dict, Any
 
 
 def _slope_between_two_points(y2: float, y1: float, x2: float, x1: float) -> float:
@@ -105,7 +105,7 @@ def _density_in_tile(x_rect: list, y_rect: list, density_profile: np.floating, n
     )
 
 
-def _is_inside_tile(x_rect: list, y_rect: list, points: np.ndarray) -> List[bool]:
+def _is_inside_tile(x_rect: list, y_rect: list, points: np.ndarray) -> np.ndarray[Any, Any]:
     polygon_tile = [[x_rect[i], y_rect[i]] for i in range(len(x_rect))]
     poly = matplotlib.path.Path(polygon_tile)
     return poly.contains_points(points)
@@ -162,8 +162,9 @@ def _create_contour_polygon_list(
         z = contour_dict[col]
         for contour_path in col.get_paths():
             for ncp, cp in enumerate(contour_path.to_polygons()):
-                lons = cp[:, 0]
-                lats = cp[:, 1]
+                cp_array = np.array(cp)
+                lons = cp_array[:, 0]
+                lats = cp_array[:, 1]
                 new_shape = geometry.Polygon([(i[0], i[1]) for i in zip(lons, lats)])
                 if ncp == 0:
                     poly = new_shape
