### 2025-01-17 #6736 @devarops
Hi @kthyng,

We have addressed most comments. We resolved all feedback from `@jhabriel` and `@ashleefv`, except for the "state of field" section. We are currently working on the last two comments from `@david-yannick`.
😉
