### 2025-01-24 1f4694f
Format using one line per sentence and add draft comment about drones
