### 2025-01-24 IslasGECI/nerd#55 @nepito
Esto le podemos responder al Yannick:
> ntentamos incorporar el efecto del viento mediante distribuciones de probabilidad no simétricas, donde el parámetro de asimetría estaba relacionado con la magnitud de la velocidad del viento. Sin embargo, la complejidad añadida por este enfoque no justificaba la mejora en la precisión de los resultados. Por esta razón, decidimos optar por la versión más simple del modelo.
