### 2025-01-24 c3dfec1
✍ Edit drone paragraph
