### 2025-01-24 5cf0af7
✍ Edit drone paragraph
