### 2025-01-28 b6008ae
😇 Add empty section State of the Field
