### 2025-01-28 344f9e7
💩🤖  Draft first version of section State of the Field
