### 2025-02-04 85e5bb0
✍🏾 Reduce sentences length
