+++ paper/paper.md
@@ -74,13 +74,15 @@ However, these approaches are time-intensive and require extensive field data, l
 NERD (Numerical Estimation of Rodenticide Density) represents a significant advancement over traditional methods by providing a highly precise, real-time solution for calculating bait density.
 Several software tools and methodologies are currently available to estimate bait distribution for rodent eradication campaigns.
 Some probabilistic models have been proposed to simulate bait distribution, but they often fail to provide the precision and speed required during aerial campaigns.
-This innovation not only reduces the time and labor required but also enhances the overall effectiveness of eradication campaigns, particularly in challenging environments like tropical islands.
-This innovative approach eliminates the need for exhaustive field measurements and enables immediate identification of bait gaps, making it particularly suited for large-scale eradication efforts.
+This innovation not only reduces the time and labor required but also enhances the overall effectiveness of eradication campaigns.
+This innovative approach eliminates the need for exhaustive field measurements and enables immediate identification of bait gaps, making it particularly for large-scale eradication efforts.
 Traditional Geographic Information Systems (GIS) tools are often used to generate bait density maps based on in situ measurements.
-Unlike GIS, which requires extensive manual input and post-processing, NERD automates the generation of bait density maps, allowing for immediate identification of bait gaps and adjustments during aerial operations.
+NERD automates the generation of bait density maps, allowing for immediate identification of bait gaps and adjustments during aerial operations.
+In contrast with NERD, GIS requires extensive manual input and post-processing.
 Unlike these alternatives, NERD offers a streamlined, real-time solution by leveraging a probability density function calibrated through controlled experiments.
-While GIS-based approaches have been widely used, they are often time-consuming and rely on untested assumptions, making them less efficient for real-time decision-making during eradication campaigns.
-While there are no direct alternative software solutions specifically designed for real-time bait density estimation in rodent eradication campaigns, NERD's unique algorithm, based on a probability density function, sets it apart by offering a more efficient and accurate tool for both planning and executing large-scale eradication efforts.
+While GIS-based approaches have been widely used, they are time-consuming, making them less efficient for real-time decision-making during eradication campaigns.
+Currently, NERD in the only software solution specifically designed for real-time bait density estimation in rodent eradication campaigns.
+NERD offers an efficient and accurate tool for both planning and executing large-scale eradication efforts. NERD's algorithm is based on a probability density function.
 
 # Formulation
 
