### 2025-02-10 e0f4869
✍🏾 Add open flightline mini reference
