+++ .github/config/.wordlist.txt
@@ -12,6 +12,7 @@ endemism
 Ensenada
 eq
 Evaristo
+Flightline
 frac
 GIS
 Grupo
