+++ paper/paper.md
@@ -70,25 +70,21 @@ With the increasing relevance and accessibility of drone technology in conservat
 
 Several software tools and methodologies are currently available to estimate bait distribution for rodent eradication campaigns.
 Currently, the field of aerial rodenticide dispersal for invasive species eradication relies heavily on Geographic Information Systems (GIS) to generate ground-level bait dispersion maps.
-Traditional Geographic Information Systems (GIS) tools, such as [Open Flightline Mini](https://github.com/zero-invasive-predators/open-flightline-mini-public), are often used to generate bait density maps based on in situ measurements.
+GIS tools, such as [Open Flightline Mini](https://github.com/zero-invasive-predators/open-flightline-mini-public), are often used to generate bait density maps based on in situ measurements.
 However, these approaches are time-intensive and require extensive field data, limiting their applicability for real-time decision-making.
+The automatic generation of bait density maps reduces the time and labor required but also enhances the overall effectiveness of eradication campaigns.
 
-GIS requires extensive manual input and post-processing.
-GIS tools requires in situ measurements to generate bait density maps.
-GIS-based approaches are time-consuming, making them less efficient for real-time decision-making during eradication campaigns.
-Geographic Information Systems (GIS) is an inefficient method to generate bait dispersion maps for aerial rodenticide dispersal for invasive species eradication.
 NERD reduces the time and labor required compared with GIS-based approach.
-
+NERD offers a streamlined, real-time solution by leveraging a probability density function calibrated through controlled experiments.
 Some probabilistic models have been proposed to simulate bait distribution, but they often fail to provide the precision and speed required during aerial campaigns.
-Unlike these alternatives, NERD offers a streamlined, real-time solution by leveraging a probability density function calibrated through controlled experiments.
-NERD offers an efficient and accurate tool for both planning and executing large-scale eradication efforts.
-Despite these advantages, further advancements, such as incorporating environmental factors like wind, remain an area for future development.
+NERD automates the generation of bait density maps, allowing for immediate identification of bait gaps and adjustments during aerial operations.
+The automatic generation of bait density maps reduces the time and labor required but also enhances the overall effectiveness of eradication campaigns.
 
 NERD represents a significant advancement over traditional methods.
-NERD automates the generation of bait density maps, allowing for immediate identification of bait gaps and adjustments during aerial operations.
-This innovative approach eliminates the need for exhaustive field measurements and enables immediate identification of bait gaps, making it particularly for large-scale eradication efforts.
-The automatic generation of bait density maps reduces the time and labor required but also enhances the overall effectiveness of eradication campaigns.
+This innovative approach eliminates the need for exhaustive field measurements and enables immediate identification of bait gaps.
+NERD offers an efficient and accurate tool for both planning and executing large-scale eradication efforts.
 Currently, NERD is the only software solution specifically designed for real-time bait density estimation in rodent eradication campaigns.
+Despite these advantages, further advancements, such as incorporating environmental factors like wind, remain an area for future development.
 
 # Formulation
 
