### 2025-02-17 IslasGECI/nerd#55 @devarops
Dear `@david-yannick`,

We appreciate your feedback provided for our article. We have addressed each of the comments and implemented the necessary changes to improve our work.

We have now uploaded the revised version and are ready for the next round of reviews if needed. Please let us know if there are any additional comments or adjustments required.

---

- We clarified that the effect of wind on bait distribution is beyond the scope of the current version [79, 87-88].
- A State of the Field section has been added [59-80].
- In the Statement of Need section, we now discuss the applicability to drones [53-56].
- Additionally, we have included [guidelines](https://github.com/IslasGECI/nerd/blob/develop/CONTRIBUTING.md) for third-party contributions.

---

Thank you for your time and dedication throughout this process. We look forward to your feedback.
