### 2025-02-17 IslasGECI/nerd#54 @devarops
Dear `@jhabriel`,

We appreciate your feedback provided for our article. We have addressed each of the comments and implemented the necessary changes to improve our work.

We have now uploaded the revised version and are ready for the next round of reviews if needed. Please let us know if there are any additional comments or adjustments required.

---

- We added docstrings and type hints to all public classes, functions and methods:
    - `nerd/calibration/best_density_function/get_best_density_function()`
    - `nerd/calibration/fit_flow_rate/fit_flow_ratefit_flow_rate()`
    - `nerd/calibration/get_swath_width/get_swath_width()`
    - `nerd/calibration/model/model()`
    - `nerd/density_functions/density_functions/uniform()`
    - `nerd/density_functions/density_functions/triangular()`
    - `nerd/density_functions/density_functions/normal()`
    - `nerd/io/Nerd/Nerd()`
    - `nerd/solver/solver()`
- We added a [citation file](https://github.com/IslasGECI/nerd/blob/develop/CITATION.cff)
- In the README file, we add the instructions for installing NERD from source
- We added a [code of conduct](https://github.com/IslasGECI/nerd/blob/develop/CODE_OF_CONDUCT.md)
- We added [contributing guidelines](https://github.com/IslasGECI/nerd/blob/develop/CONTRIBUTING.md)
- We added a section detailing how to [report bugs using Github's issues](https://github.com/IslasGECI/nerd/blob/develop/CONTRIBUTING.md#report-bugs-using-githubs-issues)
- A State of the Field section has been added [59-80]
- We addressed all your comments about the paper

---
