### 2025-02-17 #6736 @devarops
Dear @jhabriel,

Thank you for your valuable feedback on our article. We have carefully addressed all your comments and made the necessary improvements.

We have uploaded the revised version and are ready for the next round of reviews if needed. Please let us know if any further adjustments are required.

### Summary of Changes:
- Added docstrings and type hints to all public classes, functions, and methods, including:
    - `nerd/calibration/best_density_function/get_best_density_function()`
    - `nerd/calibration/fit_flow_rate/fit_flow_ratefit_flow_rate()`
    - `nerd/calibration/get_swath_width/get_swath_width()`
    - `nerd/calibration/model/model()`
    - `nerd/density_functions/density_functions/uniform()`
    - `nerd/density_functions/density_functions/triangular()`
    - `nerd/density_functions/density_functions/normal()`
    - `nerd/io/Nerd/Nerd()`
    - `nerd/solver/solver()`
- Added a [citation file](https://github.com/IslasGECI/nerd/blob/develop/CITATION.cff).
- Updated the README with installation instructions for NERD from source.
- Added a [Code of Conduct](https://github.com/IslasGECI/nerd/blob/develop/CODE_OF_CONDUCT.md).
- Included [contributing guidelines](https://github.com/IslasGECI/nerd/blob/develop/CONTRIBUTING.md).
- Added a section on [reporting bugs via GitHub issues](https://github.com/IslasGECI/nerd/blob/develop/CONTRIBUTING.md#report-bugs-using-githubs-issues).
- Included a **State of the Field** section [pp. 59-80].
- Addressed all your comments on the paper.

---

Looking forward to your feedback.
