### 2025-02-17 IslasGECI/nerd#53 @devarops
Dear `@ashleefv`,

We appreciate your feedback provided for our article. We have addressed each of the comments and implemented the necessary changes to improve our work.

We have now uploaded the revised version and are ready for the next round of reviews if needed. Please let us know if there are any additional comments or adjustments required.

---

- Included [contributing guidelines](https://github.com/IslasGECI/nerd/blob/develop/CONTRIBUTING.md).
- Included a **State of the Field** section [pp. 59-80].
- Addressed all your comments on the paper (we have fixed the typos and
have removed the code from the `paper.md` file)
- We have removed, from the README file, the broken links to Jupyter notebooks.
- Also, we have added instructions to deploy Jupyter on one's local machine using Docker. In addition, during the review process the example notebook can be inspected at http://20.36.29.86:8080
- We have removed `example_notebooks/proceedings.ipynb`.
- We have addressed all your suggestions about `example_notebooks/calibration-demo.ipynb`
- We have addressed all your suggestions about `examples/tiling_demo.ipynb`
- As suggested, we followed the PEP 8 standard that explicitly recommends absolute imports.

---

Thank you for your time and dedication throughout this process. We look forward to your feedback.
