### 2025-02-18 #6736 @jhabriel
Great work @devarops. NERD looks really good, congratulations.

All my suggestions and recommendations were properly addressed. No more comments from my side.
