### 2025-02-18 IslasGECI/nerd#53 @devarops
Dear `@ashleefv`,

We appreciate your feedback on our article. We have carefully addressed each of your comments and implemented the necessary changes to improve our work.

The revised version has now been uploaded, and we are ready for the next round of reviews if needed. Please let us know if there are any additional comments or adjustments required.

---

The following modifications have been made based on your suggestions:

- Included the [contributing guidelines](https://github.com/IslasGECI/nerd/blob/develop/CONTRIBUTING.md).
- Added a State of the Field section (lines: 59–80).
- Addressed all comments on the `paper.md`:
    - Fixed typos.
    - Removed the code from the paper.md file.
- Updated the README file:
    - Removed broken links to Jupyter notebooks.
    - Added instructions for deploying Jupyter locally using Docker. In addition, during the review process, the example notebook can be inspected at: http://20.36.29.86:8080.
- Removed `example_notebooks/proceedings.ipynb`.
- Revised example notebooks:
    - Addressed all suggestions for `example_notebooks/calibration-demo.ipynb`.
    - Addressed all suggestions for `examples/tiling_demo.ipynb`.
- Followed the PEP 8 standard, as recommended, by using absolute imports.

---

Thank you for your time and dedication throughout this process. We look forward to your feedback.
