### 2025-02-25 #6736 @jhabriel
@kthyng: My checklist is ready:

https://github.com/openjournals/joss-reviews/issues/6736#issuecomment-2154953227
