+++ paper/paper.md
@@ -18,7 +18,7 @@ authors:
     orcid: 0000-0002-3467-0008
     affiliation: 1
 affiliations:
-  - name: Grupo de Ecología y Conservación de Islas, Ensenada, Mexico
+  - name: Grupo de Ecología y Conservación de Islas
     index: 1
 bibliography: references.bib
 
