### 2025-02-28 1f9dbb6
🔖 Update release date and version in citation file
