+++ CITATION.cff
@@ -11,4 +11,6 @@ authors:
     given-names: Federico A.
     orcid: https://orcid.org/0000-0002-3467-0008
 title: "NERD: Numerical Estimation of Rodenticide Density. Journal of Open Source Software"
-version: 0.4.1
+version: 1.0.0
+date-released: 2025-02-28
+url: "https://github.com/IslasGECI/nerd"
