### 2025-02-28 #6736 @devarops
Thank you, @kthyng!

We have completed the items on the **_Additional Author Tasks After Review is Complete_** list. Here is the relevant information.

- We have double-checked the authors and affiliations, including ORCIDs.
- We have released the software with the latest changes from the review. The version number is [1.0.0](https://github.com/IslasGECI/nerd/releases/tag/v1.0.0).
- We have archived the release on Zenodo. The DOI is [10.5281/zenodo.14947594](https://doi.org/10.5281/zenodo.14947594).
- We have ensured that the title and author list, including ORCIDs, in the archive match those in the JOSS paper.
- We have verified that the license listed for the archive is consistent with the software license.
