### 2025-02-28 c0ecfff
Add DOI to CITATION.cff
