### 2025-02-28 9386285
📜 Add abstract to citation file
