+++ CHANGELOG.md
@@ -18,8 +18,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.4.1] - 2024-10-04
 
-## [0.4.0] - 2023-11-14
+## [0.4.0] - 2023-05-15
 
-[unreleased]: https://github.com/IslasGECI/nerd/compare/v0.4.0...HEAD
-[0.4.0]: https://github.com/IslasGECI/nerd/compare/v0.3.0...v0.4.0
-[0.0.1]: https://github.com/IslasGECI/nerd/releases/tag/v0.0.1
+## [0.3.1] - 2022-03-03
+
+## [0.3.0] - 2022-02-04
+
+## [0.2.0] - 2022-02-02
+
+## [0.1.0] - 2020-12-07
+
+[unreleased]: https://github.com/IslasGECI/nerd/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/IslasGECI/nerd/compare/v0.4.1...v1.0.0
+[1.0.0]: https://github.com/IslasGECI/nerd/compare/v0.4.0...v0.4.1
+[0.4.0]: https://github.com/IslasGECI/nerd/compare/v0.3.1...v0.4.0
+[0.3.1]: https://github.com/IslasGECI/nerd/compare/v0.3.0...v0.3.1
+[0.3.0]: https://github.com/IslasGECI/nerd/compare/v0.2.0...v0.3.0
+[0.2.0]: https://github.com/IslasGECI/nerd/compare/v0.1.0...v0.2.0
+[0.1.0]: https://github.com/IslasGECI/nerd/releases/tag/v0.1.0
