+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [1.0.0] - 2025-02-28
+
 ## [0.4.1] - 2024-10-04
 
 ## [0.4.0] - 2023-05-15
@@ -30,7 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [unreleased]: https://github.com/IslasGECI/nerd/compare/v1.0.0...HEAD
 [1.0.0]: https://github.com/IslasGECI/nerd/compare/v0.4.1...v1.0.0
-[1.0.0]: https://github.com/IslasGECI/nerd/compare/v0.4.0...v0.4.1
+[0.4.1]: https://github.com/IslasGECI/nerd/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/IslasGECI/nerd/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/IslasGECI/nerd/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/IslasGECI/nerd/compare/v0.2.0...v0.3.0
