+++ nerd/__init__.py
@@ -1,4 +1,4 @@
 """The NERD algorithm performs calculations with increased accuracy, displaying results almost in real-time."""
 
-__version__ = "0.4.1"
+__version__ = "1.0.0"
 from .solver import solver  # noqa
