+++ CHANGELOG.md
@@ -10,15 +10,33 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-### Fixed
-
 ### Changed
 
 ### Removed
 
+### Fixed
+
 ## [1.0.0] - 2025-02-28
+### Changed
+- Add type hint to `_is_inside_tile()`
 
 ## [0.4.1] - 2024-10-04
+### Added
+- Added docstrings and type hints to all public classes, functions, and methods, including:
+  - `nerd/calibration/best_density_function/get_best_density_function()`
+  - `nerd/calibration/fit_flow_rate/fit_flow_ratefit_flow_rate()`
+  - `nerd/calibration/get_swath_width/get_swath_width()`
+  - `nerd/calibration/model/model()`
+  - `nerd/density_functions/density_functions/uniform()`
+  - `nerd/density_functions/density_functions/triangular()`
+  - `nerd/density_functions/density_functions/normal()`
+  - `nerd/io/Nerd/Nerd()`
+  - `nerd/solver/solver()`
+### Changed
+- All the low level functions are private
+
+### Fixed
+- Followed the PEP 8 standard, as recommended, by using absolute imports.
 
 ## [0.4.0] - 2023-05-15
 
