### 2025-03-03 #6736 @kthyng
Congratulations on your new publication @devarops! Many thanks to reviewers @ashleefv, @jhabriel, and @david-yannick for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
